% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_conocimiento_region}
\alias{analizar_conocimiento_region}
\title{Title}
\usage{
analizar_conocimiento_region(llave, variable, respuesta, diseno, diccionario)
}
\arguments{
\item{diccionario}{}
}
\description{
Title
}
