% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_conocimientoRegion}
\alias{analizar_conocimientoRegion}
\title{Analizar conocimiento de personajes por region}
\usage{
analizar_conocimientoRegion(
  patron_llaveConocimiento,
  aspectos_llaveConocimiento,
  filtro_respuestaConocimiento,
  diseno,
  diccionario
)
}
\arguments{
\item{patron_llaveConocimiento}{Patrón que comparten las variables asociadas al conocimiento de un personaje}

\item{aspectos_llaveConocimiento}{Cadena de texto que diferencia las variables a analizar. Separada de patron_pregunta por un guion bajo.}

\item{filtro_respuestaConocimiento}{Filtro aplicado a las variables asociadas al conocimiento}

\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{diccionario}{Cuestionario de la encuesta en formato de procesamiento requerido}
}
\description{
Analizar conocimiento de personajes por region
}
\examples{
bd_analizar_conocimientoRegion <- analizar_conocimientoRegion(patron_llaveConocimiento = "conocimiento", aspectos_llaveConocimiento = candidatos, filtro_respuestaConocimiento = "Sí", diseno = Preguntas$Regiones$diseno, diccionario = cuestionario_demo)
}
