% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_cruceBrechas}
\alias{analizar_cruceBrechas}
\title{Analizar cruce entre dos variables}
\usage{
analizar_cruceBrechas(diseno, var1, var2_filtro, filtro, vartype)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{var2_filtro}{Variable secundaria para hacer análisis con la primaria}

\item{filtro}{Filtro aplicable a la variable secundaria}

\item{vartype}{}

\item{cruce}{Variable principal por la cual hacer análisis}
}
\description{
Analizar cruce entre dos variables
}
\examples{
analizar_cruceBrechas(diseno = diseno, var1 = "AMAI_factor", var2_filtro = "candidato_preferencia", filtro = c("Sasil de León",  "Eduardo Ramírez Aguilar"), vartype = "cv")
}
