% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_crucePuntos}
\alias{analizar_crucePuntos}
\title{Analizar cruce por puntos}
\usage{
analizar_crucePuntos(diseno, cruce, variables, vartype, valor_variables)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{cruce}{Variable principal por la cual hacer análisis}

\item{variables}{Variables secundarias para hacer análisis con la primaria}

\item{valor_variables}{Filtro aplicado a las variables secundarias}
}
\description{
Analizar cruce por puntos
}
\examples{
analizar_crucePuntos(diseno = diseno, cruce = "rurub", variables = c("conocimiento_era", "conocimiento_sasil"), vartype = "cv", valor_variables = "Sí")
}
