% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencia_multirespuesta}
\alias{analizar_frecuencia_multirespuesta}
\title{Title}
\usage{
analizar_frecuencia_multirespuesta(diseno, patron_inicial)
}
\arguments{
\item{patron_inicial}{}
}
\description{
Title
}
