% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias}
\alias{analizar_frecuencias}
\title{Analizar frecuencias de una variable}
\usage{
analizar_frecuencias(diseno, pregunta)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas}

\item{pregunta}{Nombre de la variable a calcular la estimación de proporciones de valores en la base de datos}
}
\value{
Tibble con la media de las estimaciones por valor unico de la variable seleccionada
}
\description{
Calcula la media de un diseno muestral construido con la libreria survey
}
\examples{
encuestar:::analizar_frecuencias(diseno = encuesta_demo$muestra$diseno, pregunta = "sexo")
}
