% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuenciasRegion}
\alias{analizar_frecuenciasRegion}
\title{Analizar estimaciones frecuentistas por región}
\usage{
analizar_frecuenciasRegion(regiones, variable, diseno)
}
\arguments{
\item{regiones}{Shape file de las regiones-estrato de la encuesta}

\item{variable}{Variable a obtener la estimación estimación por región}

\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}
}
\description{
Analizar estimaciones frecuentistas por región
}
\examples{
analizar_frecuenciasRegion(regiones = shp, variable = "voto_partido", diseno = diseno)
}
