% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_morena}
\alias{analizar_morena}
\title{Analizar metodología de MORENA (EN DESUSO)}
\usage{
analizar_morena(diseno, diccionario, personajes, atributos)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{diccionario}{Cuestionario de la encuesta en formato de procesamiento requerido}

\item{personajes}{Vector de nombres cortos asociados a uno o más personajes sobre los cuáles se preguntó la batería de MORENA}

\item{atributos}{Tibble que contiene los atributos y los puntos que obtiene el ganador de cada atributo}
}
\description{
Analizar metodología de MORENA (EN DESUSO)
}
\examples{
analizar_morena(diseno = diseno, diccionario = diccionario, personajes = c("era", "sasil"), atributos = atributos)
analizar_morena(diseno = diseno, diccionario = diccionario, personajes = c("era", "sasil", "jaac"), atributos = atributos)
}
