% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_respuestaAbierta}
\alias{analizar_respuestaAbierta}
\title{Analizar respuestas abiertas}
\usage{
analizar_respuestaAbierta(
  bd,
  variable,
  palabrasVacias,
  totalPalabras,
  colores = c("green", "blue", "yellow")
)
}
\arguments{
\item{bd}{Base de datos en formato largo a analizar en formato largo.}

\item{variable}{Variable cuyos valores son cadenas de texto.}

\item{palabrasVacias}{Conjunto de palabras a omitir del lenguajes español.}

\item{totalPalabras}{Total de palabras a segregar}

\item{colores}{Vector de colores en orden descendente de relevancia.}
}
\description{
Analizar respuestas abiertas
}
\examples{
analizar_respuestaAbierta(bd, variable = opinion_amlo, palabrasVacias = c("muy", "de"), totalPalabras = 15, colores = c("green", "blue", "yellow"))
}
