% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_saldo_region}
\alias{analizar_saldo_region}
\title{Title}
\usage{
analizar_saldo_region(
  llave_opinion,
  candidatos,
  ns_nc,
  cat_negativo,
  cat_regular,
  cat_positivo,
  diseno,
  diccionario
)
}
\arguments{
\item{cat_positivo}{}
}
\description{
Title
}
