% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_saldoOpinion}
\alias{analizar_saldoOpinion}
\title{Calcular el saldo de opinión por personaje}
\usage{
analizar_saldoOpinion(
  diseno,
  diccionario,
  llave_opinion,
  candidatos,
  grupo_positivo,
  grupo_negativo
)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{diccionario}{Cuestionario de la encuesta en formato de procesamiento requerido}

\item{llave_opinion}{Patrón que comparten las variables asociadas a las preguntas sobre la opinión pública hacia un personaje.}

\item{candidatos}{Cadena de texto que diferencia las variables a analizar. Separada de patron_pregunta por un guion bajo.}

\item{grupo_positivo}{Conjunto de valores de la variable opinión tratados como positivos.}

\item{grupo_negativo}{Conjunto de valores de la variable opinión tratados como negativos}
}
\description{
Calcular el saldo de opinión por personaje
}
\examples{
calcular_saldoOpinion(bd, llave_opinion = "opinion", grupo_positivo = "Buena", grupo_negativo = "Mala")
calcular_saldoOpinion(bd, llave_opinion = "op", grupo_positivo = c("Buena", "Muy buena"), grupo_negativo = c("Mala", "Muy mala"))
}
