% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_saldoRegion}
\alias{analizar_saldoRegion}
\title{Analizar saldo de opinión de personajes en cada región o estrato}
\usage{
analizar_saldoRegion(
  patron_llaveOpinion,
  aspectos_llaveOpinion,
  ns_nc,
  cat_negativo,
  cat_regular,
  cat_positivo,
  diseno,
  diccionario
)
}
\arguments{
\item{aspectos_llaveOpinion}{Cadena de texto que diferencia las variables a analizar. Separada de patron_pregunta por un guion bajo.}

\item{ns_nc}{Cadena de texto que identifica los valores relacionados al "No sabe" o "No contesta"}

\item{cat_negativo}{Vector que contiene los valores relacionados a las opiniones 'negativas'}

\item{cat_regular}{Vector que contiene los valores relacionados a las opiniones 'regulares' o neutras}

\item{cat_positivo}{Vector que contiene los valores relacionados a las opiniones 'positivas'}

\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{diccionario}{Cuestionario de la encuesta en formato de procesamiento requerido}

\item{patron_llaveConocimiento}{Patrón que comparten las variables asociadas a la opinión de un personaje}
}
\description{
Analizar saldo de opinión de personajes en cada región o estrato
}
\examples{
analizar_saldoRegion(patron_llaveOpinion = "opinion", aspectos_llaveOpinion = candidatos, ns_nc = "Ns/Nc", cat_negativo = c("Muy mala", "Mala"), cat_regular = "Regular", cat_positivo = c("Buena", "Muy buena"), diseno = Preguntas$Regiones$diseno, diccionario = Preguntas$Regiones$diccionario)
}
