% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_sankey}
\alias{analizar_sankey}
\title{Calcular datos necesarios para la funcion \code{\link[=graficar_sankey]{graficar_sankey()}}}
\usage{
analizar_sankey(diseno, variables, filtro_var1, filtro_var2)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{variables}{Vector tipo caracter que contiene los nombres de las variables de las cuales
se va a hacer el cruce}

\item{filtro_var1}{Valores de interes de la primer variable presente en el vector}

\item{filtro_var2}{Valores de interes de la primer variable presente en el vector}
}
\value{
\code{\link[tibble:tibble]{tibble()}} con las estimaciones de frecuencia para cada valor unico entre las variables
}
\description{
Calcular datos necesarios para la funcion \code{\link[=graficar_sankey]{graficar_sankey()}}
}
