% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_preguntasAbiertas.R
\name{asignar_coloresCategorias}
\alias{asignar_coloresCategorias}
\title{Title}
\usage{
asignar_coloresCategorias(
  bd_proporcionesCategorias,
  criterio = NULL,
  cuantiles = 4,
  top = 3,
  colores = c("#6a104d", "#ff8cf2")
)
}
\description{
Title
}
