% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_ganadorRegion}
\alias{calcular_ganadorRegion}
\title{Calcular valor más frecuente por región de acuerdo al diseño muestral}
\usage{
calcular_ganadorRegion(diseno, regiones, variable, lugar)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{regiones}{Shape file de las regiones-estrato de la encuesta}

\item{variable}{Variable a obtener la estimación estimación por región}

\item{lugar}{Lugar del top de las estimaciones por región}
}
\description{
Calcular valor más frecuente por región de acuerdo al diseño muestral
}
\examples{
calcular_ganadorRegion(regiones = shp, variable = "voto_partido", lugar = lugar, diseno = diseno)
}
