% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opinometro.R
\name{calcular_intentosEfectivos_opinometro}
\alias{calcular_intentosEfectivos_opinometro}
\title{Title}
\usage{
calcular_intentosEfectivos_opinometro(bd_respuestasOpinometro)
}
\arguments{
\item{bd_respuestasOpinometro}{}
}
\description{
Title
}
