% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_proporciones_nube}
\alias{calcular_proporciones_nube}
\title{Calcular proporciones por categorias de respuestas abiartas}
\usage{
calcular_proporciones_nube(
  diseno,
  llave_categorias,
  separacion_multicategoria = ">>>"
)
}
\arguments{
\item{llave_categorias}{Nombre de la variable que contiene las categorias de respuestas abiertas}

\item{separacion_multicategoria}{Cadena de texto que separa las categorias multiples}

\item{disenoDiseno}{muestral que contiene los pesos por individuo y las variables relacionadas.}
}
\description{
Calcular proporciones por categorias de respuestas abiartas
}
