% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_saldoOpinion}
\alias{calcular_saldoOpinion}
\title{Calcular el saldo de opinión por personaje}
\usage{
calcular_saldoOpinion(bd, llave_opinion, grupo_positivo, grupo_negativo)
}
\arguments{
\item{bd}{Base de datos producto de 'analizar_frecuencias_aspectos'.}

\item{llave_opinion}{Patrón que comparten las variables asociadas a las preguntas sobre la opinión pública hacia un personaje.}

\item{grupo_positivo}{Conjunto de valores de la variable opinión tratados como positivos.}

\item{grupo_negativo}{Conjunto de valores de la variable opinión tratados como negativos}
}
\description{
Calcular el saldo de opinión por personaje
}
\examples{
calcular_saldoOpinion(bd, llave_opinion = "opinion", grupo_positivo = "Buena", grupo_negativo = "Mala")
calcular_saldoOpinion(bd, llave_opinion = "op", grupo_positivo = c("Buena", "Muy buena"), grupo_negativo = c("Mala", "Muy mala"))
}
