% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_tasa_rechazo}
\alias{calcular_tasa_rechazo}
\title{Calcular el porcentaje de intentos rechazados de inciar una encuesta}
\usage{
calcular_tasa_rechazo(bd_respuestas_efectivas, por_usuario = FALSE)
}
\arguments{
\item{bd_respuestas_efectivas}{\code{\link[tibble:tibble]{tibble::tibble()}} que contiene las respuesyas efectivas}

\item{usuario_plataforma}{Logical. Para determinar si el calculo se hace por usuario de plataforma o por
total de respuestas}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} con variable rechazo en escala de porcentaje natural
}
\description{
Calcular el porcentaje de intentos rechazados de inciar una encuesta
}
