% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checar_formato_surveytogo.R
\name{checar_formato_diccionario_surveytogo}
\alias{checar_formato_diccionario_surveytogo}
\title{Title}
\usage{
checar_formato_diccionario_surveytogo(encuesta)
}
\arguments{
\item{encuesta}{}
}
\value{

}
\description{
Title
}
