% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opinometro.R
\name{determinar_contenidoCuestionario}
\alias{determinar_contenidoCuestionario}
\title{Title}
\usage{
determinar_contenidoCuestionario(pool, id_cuestionario)
}
\arguments{
\item{id_cuestionario}{}
}
\description{
Title
}
