% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opinometro.R
\name{determinarVariables_cuestinoarioOpinometro}
\alias{determinarVariables_cuestinoarioOpinometro}
\title{Title}
\usage{
determinarVariables_cuestinoarioOpinometro(pool, id_cuestionario)
}
\arguments{
\item{id_cuestionario}{}
}
\description{
Title
}
