% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clase_resultados.R
\name{Especial}
\alias{Especial}
\title{Clase Especial}
\description{
La clase \code{Especial} contiene los métodos necesarios para producir los resultados
cuya naturaleza inherente de la pregunta requiera de una adaptación particular o compuesta
para visualizar los resultados.
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{encuesta}}{Clase de jerarquía superior. Se requiere para extraer el diseno de la encuesta.}

    \item{\code{diseno}}{Objeto tipo \link{design} de la paquetería \link{survey} relacionado con el diseno de la
encuesta.}

    \item{\code{diccionario}}{Diccionario de procesamiento de la encuesta.}

    \item{\code{tema}}{Objeto tipo \link[ggplot2:theme]{theme} de la paquetería \link[ggplot2:ggplot2]{ggplot2}}

    \item{\code{graficadas}}{Campo heredado por la clase \code{Resultados}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Especial-initialize}{\code{Especial$new()}}
    \item \href{#method-Especial-candidatoOpinion}{\code{Especial$candidatoOpinion()}}
    \item \href{#method-Especial-candidatoOpinion2}{\code{Especial$candidatoOpinion2()}}
    \item \href{#method-Especial-candidatoPartido}{\code{Especial$candidatoPartido()}}
    \item \href{#method-Especial-candidatoSaldo}{\code{Especial$candidatoSaldo()}}
    \item \href{#method-Especial-metodo_morena}{\code{Especial$metodo_morena()}}
    \item \href{#method-Especial-clone}{\code{Especial$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-initialize}{}}}
\subsection{\code{Especial$new()}}{
  Define los insumos necesarios para preparar los métodos que producen los resultados
que requieren un tipo de visualización particular o compuesta.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{encuesta}}{Clase de jerarquía superior. Es el formato de la paquetería \code{encuestar}. Contiene
el diseno muestral de la encuesta.}
      \item{\code{diseno}}{Objeto tipo \link{design} de la paquetería \link{survey}. No se espera que los parametros
\code{encuesta} y \code{diseno} sean no nullos. El segundo sobreescribe el primero.}
      \item{\code{diccionario}}{Diccionario (o codebook) del cuestionario de la encuesta.}
      \item{\code{tema}}{Objeto tipo \link[ggplot2:theme]{theme} de la paquetería \link[ggplot2:ggplot2]{ggplot2}.}
      \item{\code{graficadas}}{Parametro heredado de la clase \code{Resultados}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-candidatoOpinion"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-candidatoOpinion}{}}}
\subsection{\code{Especial$candidatoOpinion()}}{
  Visualiza los resultados de preguntas con mismas posibles respuestas y aplicadas
sobre diferentes tópicos. Usualmente este tipo de preguntas son asociadas a la opinión y/o
el conocimiento de personajes. El conjunto de preguntas están identificas, por ejemplo, como
conocimiento_persona1, conocimiento_persona2, etc y opinion_persona1, opinion_persona2, etc.
Por lo general, las variables de opinion destán condicionadas por las de conocimiento. Esto
está considerado en el procesamiento de los resultados.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$candidatoOpinion(
  patron_inicial,
  aspectos,
  ns_nc = "Ns/Nc",
  regular = "Regular",
  llave_burbuja = NA,
  filtro_burbuja = "respuesta == 'Sí'",
  size_burbuja = 8,
  grupo_positivo,
  grupo_negativo,
  orden_resp,
  colores_opinion = c("red", "yellow", "green"),
  color_nsnc = "gray70",
  color_burbuja = "blue",
  caption_opinion = "",
  caption_nsnc = "Ns/Nc",
  caption_burbuja = "Conocimiento",
  size_text_legend = 12,
  size_caption_opinion = 12,
  size_caption_nsnc = 14,
  size_caption_burbuja = 14,
  size_text_cat = 16,
  size_pct = 12,
  burbuja = T,
  salto = 20,
  salto_respuestas = 100,
  mostrar_nsnc = T,
  orden_cat = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{patron_inicial}}{Valor tipo caracter. Es el patron inicial de caracteres que comparten
los nombres de las variables que representan la opinion. En el ejemplo este parámetro seía
"opinion".}
      \item{\code{aspectos}}{Vector tipo caracter que identifica a las diferentes variables a procesar. De
acuerdo al método, el ejemplo sería c("persona1", "persona2").}
      \item{\code{ns_nc}}{Valor tipo caracter. Es el valor asociado a la respuesta  'No sabe o no contesta'
en la pregunta de opinión. Ya que este valor no aporta información, se separa de los resultados.}
      \item{\code{regular}}{Valor tipo caracter. Es el valor asociado a la respuesta  'Regular' que
representa una opinión neutra.}
      \item{\code{llave_burbuja}}{Valor tipo caracter. Si es no vacío, se visualizan las estimaciones
relacionadas a la pregunta de conocimiento del personaje.}
      \item{\code{filtro_burbuja}}{Valor tipo caracter. Operación de identidad que se aplica como filtro
entre los resultados de las preguntas asociadas al conocimiento. Por lo general el filtro
tiene la forma filtro = "respuesta == 'Sí'". Este parámetro no afecta al cálculo de las
estimaciones de las variables asociadas a la opinion, unicamente a las asociadas al conocimiento.}
      \item{\code{size_burbuja}}{Valor tipo entero. Determina el tamano de las "burbujas" que se usan para
visualizar los resultados de las preguntas asociadas al conocimiento. Es el parámetro \link{max_size}
de la función \code{\link[ggplot2:scale_size_area]{ggplot2::scale_size_area()}}.}
      \item{\code{grupo_positivo}}{Vector tipo caracter. Indica los valores únicos de las variables asociadas
a la pregunta de opinión que se consideran como 'opinion positiva" con el objetivo de ordenarlas
de acuerdo a un gradiente de opinión}
      \item{\code{grupo_negativo}}{Vector tipo caracter. Indica los valores únicos de las variables asociadas
a la pregunta de opinión que se consideran como 'opinion negativa" con el objetivo de ordenarlas
de acuerdo a un gradiente de opinión}
      \item{\code{orden_resp}}{Vector tipo caracter que contiene los valores posibles de las respuestas
de preguntas asociadas a la opinión.}
      \item{\code{colores_opinion}}{Vector tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena'cada proporción de opinión.}
      \item{\code{color_nsnc}}{Valor tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena' la proporción asociada a la respuesta 'No sabe o no contesta".}
      \item{\code{color_burbuja}}{Valor tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena' el gráfico relacionado a los resultados de la pregunta de conocimiento.}
      \item{\code{caption_burbuja}}{Valor tipo entero. Determina el tamano del texto mostrado en el
\link{caption} de la visualizacion relacionada al conocimiento de la persona. Es el
parámetro \link[ggplot2:size]{size} de la función \code{\link[ggplot2:element_text]{ggplot2::element_text()}} aplicada al argumento
\link{plot.caption} del \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
      \item{\code{size_caption_opinion}}{Valor tipo entero. Determina el tamano del texto mostrado en el
\link{caption} de la visualizacion principal. Es el parámetro \link[ggplot2:size]{size} de la función
\code{\link[ggplot2:element_text]{ggplot2::element_text()}} aplicada al argumento \link{plot.caption} del \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
      \item{\code{size_caption_nsnc}}{Valor tipo entero. Determina el tamano del texto mostrado en el
\link{caption} de la visualizacion relacionada a la respuesta 'No sabe o no contesta". Es el
parámetro \link[ggplot2:size]{size} de la función \code{\link[ggplot2:element_text]{ggplot2::element_text()}} aplicada al argumento
\link{plot.caption} del \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
      \item{\code{size_text_cat}}{Valor tipo entero. Determina el tamano del texto mostrado en las
etiquetas relacionadas a la opinión de personajes. Es el parámetro \link[ggplot2:size]{size} de la función
\code{\link[ggplot2:element_text]{ggplot2::element_text()}} aplicada al argumento \link{axis.text.y} del \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
      \item{\code{size_pct}}{Valor tipo entero. Determina el tamano del texto mostrado dentro de las
barras que representan las proporciones de opinion. Es el parámetro \link[ggplot2:size]{size} de la función
\code{\link[ggfittext:geom_fit_text]{ggfittext::geom_fit_text()}}.}
      \item{\code{burbuja}}{\code{LOGICAL}. Determina si mostrar o no el gráfico relacionado a los resultados
de la pregunta de conocimiento.}
      \item{\code{salto}}{Valor tipo entero. Reduce o aumenta la longitud de las categorías en el eje x.}
      \item{\code{salto_respuestas}}{Valor tipo entero. Reduce o aumenta la longitud de las categorías
posibles de respuestas de las preguntas de opinion.}
      \item{\code{mostrar_nsnc}}{\code{LOGICAL}. Determina si mostrar o no los resultados de los valores
"No sabe o no contesta".}
      \item{\code{orden_cat}}{Vector ordenado tipo caracter. Son los sufijos que las variables asociadas
a las preguntas y ordena los resultados de acuerdo a dicho vector.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-candidatoOpinion2"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-candidatoOpinion2}{}}}
\subsection{\code{Especial$candidatoOpinion2()}}{
  Visualiza los resultados de preguntas con mismas posibles respuestas y aplicadas
sobre diferentes tópicos. Usualmente este tipo de preguntas son asociadas a la opinión y/o
el conocimiento de personajes. El conjunto de preguntas están identificas, por ejemplo, como
conocimiento_persona1, conocimiento_persona2, etc y opinion_persona1, opinion_persona2, etc.
Por lo general, las variables de opinion destán condicionadas por las de conocimiento. Esto
está considerado en el procesamiento de los resultados. La vsualizacion es en forma de
tabla usando la paquetería \link{flextable}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$candidatoOpinion2(
  patron_opinion,
  aspectos,
  ns_nc = "Ns/Nc",
  regular = "Regular",
  patron_conocimiento,
  filtro_conocimiento = "respuesta == 'Sí'",
  orden_opinion,
  etiquetas = c("Candidato", "Opinión"),
  colores_opinion = c("red", "yellow", "green"),
  colores_candidato = colores_candidato,
  color_principal = "#BF498A",
  color_conocimiento = "blue",
  size_text_header = 18,
  size_text_body = 14,
  salto = 20,
  salto_respuestas = 100
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{patron_opinion}}{Valor tipo caracter. Es el patron inicial de caracteres que comparten
los nombres de las variables que representan la opinion. En el ejemplo este parámetro seía
"opinion".}
      \item{\code{aspectos}}{Vector tipo caracter que identifica a las diferentes variables a procesar. De
acuerdo al método, el ejemplo sería c("persona1", "persona2").}
      \item{\code{ns_nc}}{Valor tipo caracter. Es el valor asociado a la respuesta  'No sabe o no contesta'
en la pregunta de opinión. Ya que este valor no aporta información, se separa de los resultados.}
      \item{\code{regular}}{Valor tipo caracter. Es el valor asociado a la respuesta  'Regular' que
representa una opinión neutra.}
      \item{\code{patron_conocimiento}}{Valor tipo caracter. Es el patron inicial de caracteres que comparten
los nombres de las variables que representan el conocimiento En el ejemplo este parámetro seía
"opinion".}
      \item{\code{filtro_conocimiento}}{Valor tipo caracter. Operación de identidad que se aplica como filtro
entre los resultados de las preguntas asociadas al conocimiento. Por lo general el filtro
tiene la forma filtro = "respuesta == 'Sí'". Este parámetro no afecta al cálculo de las
estimaciones de las variables asociadas a la opinion, unicamente a las asociadas al conocimiento.}
      \item{\code{orden_opinion}}{Vector tipo caracter que contiene los valores posibles de las respuestas
de preguntas asociadas a la opinión. Las columnas de la tabla se orden de acuerdo a este vector}
      \item{\code{etiquetas}}{Dupla ordenada tipo caracter. Nombres de las variables para mostrar en el
resultado final.}
      \item{\code{colores_opinion}}{Vector tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena'cada proporción de opinión. Tantos colores como opiniones posibles.}
      \item{\code{colores_candidato}}{Vector que identifica los colores asociados a los valores únicos de la
variable relacionada al candidato.}
      \item{\code{color_principal}}{Valor tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena'la parte principal de la tabla}
      \item{\code{color_conocimiento}}{Valor tipo caracter. Color en formato \code{HEX}. Es el color con el que
se 'rellena'la columna de la tabla relacionada con los resultados de conocimiento.}
      \item{\code{size_text_header}}{Valor entero. Define el tamano del texto del encabezado}
      \item{\code{size_text_body}}{Valor entero. Define el tamano del texto del cuerpo de la tabla}
      \item{\code{salto}}{Valor entero. Define la longitud del texto de las filas}
      \item{\code{salto_respuestas}}{Valor entero. Define la longitu del texto de los encabezadps.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-candidatoPartido"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-candidatoPartido}{}}}
\subsection{\code{Especial$candidatoPartido()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$candidatoPartido(
  llave_partido,
  llave_conocimiento,
  respuesta_conoce,
  candidatos,
  corte_otro,
  cliente,
  colores_candidatos,
  colores_partido,
  corte_vis = 0,
  size_text = 6,
  size_text_conocimiento = 6
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-candidatoSaldo"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-candidatoSaldo}{}}}
\subsection{\code{Especial$candidatoSaldo()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$candidatoSaldo(
  llave_opinion,
  candidatos,
  positivos,
  negativos,
  regular = "Regular",
  ns_nc = "Ns/Nc",
  color_positivo = "green",
  color_negativo = "red",
  orden_cat = NULL,
  salto_cat = 25,
  caption_opinion,
  size_text_cat = 14,
  size_text_legend = 10,
  size_pct = 12,
  size_caption_opinion = 10
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-metodo_morena"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-metodo_morena}{}}}
\subsection{\code{Especial$metodo_morena()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$metodo_morena(personajes, atributos)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Especial-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Especial-clone}{}}}
\subsection{\code{Especial$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Especial$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
