% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{formatear_tabla_candidatoOpinion}
\alias{formatear_tabla_candidatoOpinion}
\title{Title}
\usage{
formatear_tabla_candidatoOpinion(
  tabla_candidatoOpinion,
  orden_opinion,
  etiquetas,
  colores_opinion,
  color_principal,
  colores_candidato,
  size_text_header,
  size_text_body,
  salto,
  color_conocimiento
)
}
\description{
Title
}
