% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{formatear_tabla_votoCruzado}
\alias{formatear_tabla_votoCruzado}
\title{Title}
\usage{
formatear_tabla_votoCruzado(
  tabla_votoCruzado,
  var1,
  var2,
  filtro_var2,
  etiquetas,
  colores_var1,
  colores_var2,
  size_text_header,
  size_text_body,
  salto
)
}
\description{
Title
}
