% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Grafica}
\alias{Grafica}
\title{Esta es la clase de Grafica}
\description{
Esta es la clase de Grafica

Esta es la clase de Grafica
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Grafica-new}{\code{Grafica$new()}}
\item \href{#method-Grafica-gauge_numerica}{\code{Grafica$gauge_numerica()}}
\item \href{#method-Grafica-gauge_categorica}{\code{Grafica$gauge_categorica()}}
\item \href{#method-Grafica-barras_categorica}{\code{Grafica$barras_categorica()}}
\item \href{#method-Grafica-barras_numerica}{\code{Grafica$barras_numerica()}}
\item \href{#method-Grafica-barras_multirespuesta}{\code{Grafica$barras_multirespuesta()}}
\item \href{#method-Grafica-barras_texto}{\code{Grafica$barras_texto()}}
\item \href{#method-Grafica-intervalo_numerica}{\code{Grafica$intervalo_numerica()}}
\item \href{#method-Grafica-nube_texto}{\code{Grafica$nube_texto()}}
\item \href{#method-Grafica-sankey_categorica}{\code{Grafica$sankey_categorica()}}
\item \href{#method-Grafica-candidatoOpinion}{\code{Grafica$candidatoOpinion()}}
\item \href{#method-Grafica-candidatoPartido}{\code{Grafica$candidatoPartido()}}
\item \href{#method-Grafica-candidatoSaldo}{\code{Grafica$candidatoSaldo()}}
\item \href{#method-Grafica-metodo_morena}{\code{Grafica$metodo_morena()}}
\item \href{#method-Grafica-puntos}{\code{Grafica$puntos()}}
\item \href{#method-Grafica-brechasDuales}{\code{Grafica$brechasDuales()}}
\item \href{#method-Grafica-brechasMultiples}{\code{Grafica$brechasMultiples()}}
\item \href{#method-Grafica-barrasMultiples}{\code{Grafica$barrasMultiples()}}
\item \href{#method-Grafica-bloques}{\code{Grafica$bloques()}}
\item \href{#method-Grafica-correspondencia}{\code{Grafica$correspondencia()}}
\item \href{#method-Grafica-componentesPrincipales}{\code{Grafica$componentesPrincipales()}}
\item \href{#method-Grafica-blackBox}{\code{Grafica$blackBox()}}
\item \href{#method-Grafica-clone}{\code{Grafica$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-new"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$new(diseno, diccionario = NULL, tema)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-gauge_numerica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-gauge_numerica}{}}}
\subsection{Method \code{gauge_numerica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$gauge_numerica(
  codigo,
  color = "#850D2D",
  escala = c(0, 10),
  size_text_pct = 14
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-gauge_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-gauge_categorica}{}}}
\subsection{Method \code{gauge_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$gauge_categorica(
  codigo,
  filtro,
  color = "#850D2D",
  escala = c(0, 1),
  size_text_pct = 14
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-barras_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-barras_categorica}{}}}
\subsection{Method \code{barras_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$barras_categorica(
  codigo,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-barras_numerica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-barras_numerica}{}}}
\subsection{Method \code{barras_numerica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$barras_numerica(
  patron_inicial,
  aspectos = NULL,
  salto = 20,
  filtro = NULL,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-barras_multirespuesta"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-barras_multirespuesta}{}}}
\subsection{Method \code{barras_multirespuesta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$barras_multirespuesta(
  patron_inicial,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-barras_texto"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-barras_texto}{}}}
\subsection{Method \code{barras_texto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$barras_texto()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-intervalo_numerica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-intervalo_numerica}{}}}
\subsection{Method \code{intervalo_numerica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$intervalo_numerica(
  patron,
  aspectos,
  escala = c(0, 10),
  point_size = 1,
  text_point_size = 8
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-nube_texto"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-nube_texto}{}}}
\subsection{Method \code{nube_texto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$nube_texto(
  codigo,
  palabrasVacias = NULL,
  total_palabras = 15,
  colores = c("#619CFF", "#FFFF33", "#00BA38")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-sankey_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-sankey_categorica}{}}}
\subsection{Method \code{sankey_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$sankey_categorica(variables = NULL, size_text_cat = 8)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-candidatoOpinion"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-candidatoOpinion}{}}}
\subsection{Method \code{candidatoOpinion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$candidatoOpinion(
  patron_inicial,
  aspectos,
  ns_nc = "Ns/Nc",
  regular = "Regular",
  llave_burbuja = "conocimiento",
  filtro_burbuja = "respuesta == 'Sí'",
  grupo_positivo,
  grupo_negativo,
  orden_resp,
  colores_opinion = c("red", "yellow", "green", "gray70"),
  color_burbuja = "blue",
  caption_opinion = "",
  caption_nsnc = "Ns/Nc",
  caption_burbuja = "Conocimiento",
  size_caption_opinion = 12,
  size_caption_nsnc = 14,
  size_caption_burbuja = 14,
  size_text_cat = 16,
  burbuja = T,
  salto = 20
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-candidatoPartido"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-candidatoPartido}{}}}
\subsection{Method \code{candidatoPartido()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$candidatoPartido(
  llave_partido,
  llave_conocimiento,
  respuesta_conoce,
  candidatos,
  corte_otro,
  cliente,
  colores_candidatos,
  colores_partido
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-candidatoSaldo"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-candidatoSaldo}{}}}
\subsection{Method \code{candidatoSaldo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$candidatoSaldo(
  patron_inicial,
  aspectos,
  positivos,
  negativos,
  color_positivo = "green",
  color_negativo = "red"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-metodo_morena"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-metodo_morena}{}}}
\subsection{Method \code{metodo_morena()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$metodo_morena(personajes, atributos)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-puntos"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-puntos}{}}}
\subsection{Method \code{puntos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$puntos(cruce, variables, vartype = "se", valor_variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-brechasDuales"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-brechasDuales}{}}}
\subsection{Method \code{brechasDuales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$brechasDuales(
  var1,
  var2_filtro,
  filtro,
  vartype = "cv",
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-brechasMultiples"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-brechasMultiples}{}}}
\subsection{Method \code{brechasMultiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$brechasMultiples(
  por_grupo,
  variables,
  vartype = "cv",
  valor_variables,
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-barrasMultiples"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-barrasMultiples}{}}}
\subsection{Method \code{barrasMultiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$barrasMultiples(
  por_grupo,
  variables,
  vartype = "cv",
  valor_variables,
  color = "green",
  filter = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-bloques"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-bloques}{}}}
\subsection{Method \code{bloques()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$bloques(cruce, variable, vartype = "cv", filter = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-correspondencia"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-correspondencia}{}}}
\subsection{Method \code{correspondencia()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$correspondencia(
  var1,
  var2,
  legenda1 = NULL,
  legenda2 = NULL,
  colores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-componentesPrincipales"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-componentesPrincipales}{}}}
\subsection{Method \code{componentesPrincipales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$componentesPrincipales(variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-blackBox"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-blackBox}{}}}
\subsection{Method \code{blackBox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$blackBox(vars, stimuli)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Grafica-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Grafica-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Grafica$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
