% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_aspectos_frecuencias}
\alias{graficar_aspectos_frecuencias}
\title{Title}
\usage{
graficar_aspectos_frecuencias(
  bd,
  titulo = NULL,
  fill = NULL,
  nota = "",
  grupo_positivo = c("Aprueba mucho", "Aprueba poco"),
  grupo_negativo = c("Desaprueba mucho", "Desaprueba poco"),
  ns_nc = "Ns/Nc",
  colores = c(`Desaprueba mucho` = "#DE6400", `Desaprueba poco` = "#FB8500", `Aprueba
    poco` = "#126782", `Aprueba mucho` = "#023047", `Ns/Nc` = "gray"),
  orden = c("Desaprueba poco", "Desaprueba mucho", "Aprueba poco", "Aprueba mucho"),
  familia = "Poppins"
)
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{titulo}{Título de la gráfica}

\item{fill}{Variable de color}

\item{nota}{Nota opcional}

\item{grupo_positivo}{Grupo cuyas barras serán positivas}

\item{grupo_negativo}{Grupo que su porcentaje será negativo}

\item{ns_nc}{Nombre de las respuestas que corresponden a "No sabe", default: "Ns/Nc"}

\item{colores}{Vector de colores con el nombre de las variables}

\item{orden}{Orden de la varble de color}

\item{familia}{font family}
}
\value{

}
\description{
Title
}
