% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras}
\alias{graficar_barras}
\title{Gráfica de barras horizontales}
\usage{
graficar_barras(
  bd,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0,
  orden_respuestas = NA,
  text_size = 8
)
}
\arguments{
\item{bd}{Base de datos con una variable categórica (respuesta) y una numérica (media).}

\item{salto}{Número entero, se aplica un stringr::str_wrap a la variable categórica.}

\item{porcentajes_fuera}{Si es T, las labels de los porcentajes aparecen fuera (o sobre) las barras.}

\item{desplazar_porcentajes}{Si porcentajes_fuera es T, este parametro ajusta las etiquetas de texto.}

\item{orden_respuestas}{Vector ordenado tipo caracter usado para ordenar los valores del eje x}
}
\value{
Objeto tipo \link[ggplot2:ggplot]{ggplot}
}
\description{
Recibe un \code{\link[tibble:tibble]{tibble()}} y genera un objeto tipo \link[ggplot2:ggplot]{ggplot}. El producto es una grafica de barras
ordenadas de acuerdo al criterio indicado.
}
\examples{
encuestar:::analizar_frecuencias(diseno = encuesta_demo$muestra$diseno, pregunta = "voto_pm_24") |> encuestar:::graficar_barras()
encuestar:::analizar_frecuencias(diseno = encuesta_demo$muestra$diseno, pregunta = "seguridad_sonora") |> encuestar:::graficar_barras()
}
