% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_frecuencia}
\alias{graficar_barras_frecuencia}
\title{Para graficar frecuencias simples}
\usage{
graficar_barras_frecuencia(bd, titulo, salto = 20, nota = "", tema)
}
\arguments{
\item{bd}{Debe provenir de la función de analizar_frecuencias}

\item{titulo}{Es un parámetro obligatorio para el título de la gráfica}

\item{nota}{Si de desea añadir una nota al pie del plot}

\item{familia}{Familia tipográfica de los elementos del plot}

\item{color_etiqueta}{Color de la letra de las etiquetas de datos}

\item{colores}{Vector de colores del plot}
}
\description{
Para graficar frecuencias simples
}
