% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_grupos}
\alias{graficar_barras_grupos}
\title{Para graficar frecuencias de barras agrupadas}
\usage{
graficar_barras_grupos(
  bd,
  titulo,
  familia = "Poppins",
  color_etiqueta = "#3B3838",
  nota = "",
  orden = "",
  colores
)
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{titulo}{Es un parámetro obligatorio para el título de la gráfica}

\item{familia}{Familia tipográfica de los elementos del plot}

\item{color_etiqueta}{Color de la letra de las etiquetas de datos}

\item{nota}{Si de desea añadir una nota al pie del plot}

\item{orden}{Vector en el orden que se desean los grupos de las barras}

\item{colores}{Vector de colores del plot}
}
\value{

}
\description{
Para graficar frecuencias de barras agrupadas
}
