% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_palabras}
\alias{graficar_barras_palabras}
\title{Title}
\usage{
graficar_barras_palabras(bd, pregunta, n = 10)
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{pregunta}{columna de caracter a procesar}

\item{n}{Número de palabras a mostrar}
}
\value{

}
\description{
Title
}
