% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_saldo}
\alias{graficar_barras_saldo}
\title{Graficar barras saldo}
\usage{
graficar_barras_saldo(
  bd,
  orden,
  grupo_positivo,
  grupo_negativo,
  Regular,
  colores,
  salto_respuestas,
  salto_tema,
  caption_opinion,
  size_text_cat = 12,
  size_pct,
  size_caption_opinion,
  size_text_legend = 12,
  tema = encuestar::tema_morant()
)
}
\arguments{
\item{bd}{\code{\link[=tibble]{tibble()}} que contiene las variables necesarias para generar la grafica}

\item{orden}{Vector ordenado tipo caracter usado para ordenar los valores del eje x}

\item{grupo_positivo}{Conjunto de valores de la  variable 'respuesta' considerados como positivos.}

\item{grupo_negativo}{Conjunto de valores de la  variable 'respuesta' considerados como negativos.}

\item{Regular}{Valor de la variable 'respuesta' asociado a "Regular" en la pregunta relativa.}

\item{colores}{Vector ordenado de colores asociados al grupo negativo, regular y positivo.}

\item{salto_respuestas}{Valor entero usato como parametro en \code{\link[stringr:str_wrap]{stringr::str_wrap()}} aplicado a todos las leyendas del grafico}

\item{salto_tema}{Valor entero usato como parametro en \code{\link[stringr:str_wrap]{stringr::str_wrap()}} aplicado a todos las categorias del eje x}

\item{caption_opinion}{Cadena de texto usada en el parametro \link{caption} de la funcion \code{\link[ggplot2:labs]{ggplot2::labs()}}}

\item{size_text_cat}{Parametro \link{size} de la funcion \code{\link[ggplot2:element]{ggplot2::element_text()}} usado en el parametro \link{axis.text.x} en el tema particular del grafico}

\item{size_pct}{Parametro \link{size} de la funcion \code{\link[ggfittext:geom_fit_text]{ggfittext::geom_fit_text()}} que controla el tamano del texto que muestra el porcentaje dentro de las barras}

\item{size_caption_opinion}{Parametro \link{size} de la funcion \code{\link[ggplot2:element]{ggplot2::element_text()}} usado en el parametro \link{plot.caption} en el tema particular del grafico}

\item{tema}{Tema grafico de \link{ggplot2}}
}
\value{
Objeto tipo \link{ggplot}
}
\description{
Recibe un \code{\link[=tibble]{tibble()}} y genera un objeto tipo \link{ggplot}. El producto es una grafica tipo barras horizontales
diferenciando dos grupos principales: positivos y negativos. Los resultados mostrados omiten la
categoria de No sabe No contesta
}
