% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_candidato_partido}
\alias{graficar_candidato_partido}
\title{Title}
\usage{
graficar_candidato_partido(
  bases,
  cliente,
  tipo_conoce,
  colores_candidato,
  solo_respondidos,
  colores_partido,
  tema
)
}
\arguments{
\item{bases}{}
}
\description{
Title
}
