% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_conocimientoRegion}
\alias{graficar_conocimientoRegion}
\title{Graficar conocimiento de personajes por regi칩n o estrato}
\usage{
graficar_conocimientoRegion(bd, ordenRegiones, salto_labelRegiones = 5)
}
\arguments{
\item{bd}{Base de datos resultado de la funci칩n 'analizar_conocimientoRegion'}

\item{ordenRegiones}{Vector que indica el orden de las columnas del geom_tile}

\item{salto_labelRegiones}{Par치metro 'width' usado por stringr::str_wrap en las etiquetas superiores}
}
\description{
Graficar conocimiento de personajes por regi칩n o estrato
}
