% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_barrasMultiples}
\alias{graficar_cruce_barrasMultiples}
\title{Graficar cruce de una variable vs múltiples variables unando gráficas de barras}
\usage{
graficar_cruce_barrasMultiples(bd, cruce, vartype, color, filter)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_crucePuntos'}

\item{cruce}{Variable principal por la cual se hace el cruce}

\item{color}{Color usado en el parámetro 'fill' de la función 'aes' de ggplot2}

\item{filter}{Filtro aplicable al parámetro 'cruce'}
}
\description{
Graficar cruce de una variable vs múltiples variables unando gráficas de barras
}
