% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_brechasDuales}
\alias{graficar_cruce_brechasDuales}
\title{Graficar cruce de una variable vs otra con opción a filtro}
\usage{
graficar_cruce_brechasDuales(
  bd,
  var1,
  var2_filtro,
  vartype = "cv",
  line_rich = F,
  line_linewidth = 2,
  line_hjust = 0.5,
  line_vjust = -0.5
)
}
\arguments{
\item{bd}{Base de datos resultado de la función 'analizar_cruceBrechas'}

\item{var1}{Variable principal por la cual se hace el cruce}

\item{var2_filtro}{Variable secundaria para hacer análisis con la primaria}

\item{line_rich}{Argumento de la función 'geom_textline'}

\item{line_linewidth}{Argumento de la función 'geom_textline'}

\item{line_hjust}{Argumento de la función 'geom_textline'}

\item{line_vjust}{Argumento de la función 'geom_textline'}
}
\description{
Graficar cruce de una variable vs otra con opción a filtro
}
\examples{
graficar_cruce_brechasDuales(bd, var1 = "AMAI_factor", var2_filtro = "candidato_preferencia")
}
