% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_brechasMultiples}
\alias{graficar_cruce_brechasMultiples}
\title{Graficar cruce de una variable vs múltiples variables}
\usage{
graficar_cruce_brechasMultiples(
  bd,
  cruce,
  vartype,
  line_rich,
  line_linewidth,
  line_hjust,
  line_vjust
)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_crucePuntos'}

\item{cruce}{Variable principal por la cual se hace el cruce}

\item{line_rich}{Argumento de la función 'geom_textline'}

\item{line_linewidth}{Argumento de la función 'geom_textline'}

\item{line_hjust}{Argumento de la función 'geom_textline'}

\item{line_vjust}{Argumento de la función 'geom_textline'}
}
\description{
Graficar cruce de una variable vs múltiples variables
}
