% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_puntosMultiples}
\alias{graficar_cruce_puntosMultiples}
\title{Graficar cruce de una variable vs varias variables pero filtradas}
\usage{
graficar_cruce_puntosMultiples(bd, orden_variablePrincipal)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_crucePuntos'}

\item{orden_variablePrincipal}{Factor de la variable principal}
}
\description{
Graficar cruce de una variable vs varias variables pero filtradas
}
