% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_crucePuntos}
\alias{graficar_crucePuntos}
\title{Graficar cruce de una variable vs varias variables}
\usage{
graficar_crucePuntos(bd, cruce, vartype)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_crucePuntos'}

\item{cruce}{Variable principal por la cual se hace el cruce}

\item{vartype}{}
}
\description{
Graficar cruce de una variable vs varias variables
}
\examples{
graficar_cruce_puntos(bd, cruce = "rurub", vartype = "cv")
}
