% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_estratos_grupos}
\alias{graficar_estratos_grupos}
\title{Title}
\usage{
graficar_estratos_grupos(bd, titulo = "", nota = "", familia = "Poppins")
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{familia}{}
}
\description{
Title
}
