% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_frecuencia_opuestos}
\alias{graficar_frecuencia_opuestos}
\title{Title}
\usage{
graficar_frecuencia_opuestos(
  bd,
  titulo,
  grupo1,
  grupo2,
  color1 = "#006466",
  color2 = "#4d194d",
  contraste_etiqueta = T,
  color_etiqueta = "#FFFFFF"
)
}
\arguments{
\item{titulo}{Es un parámetro obligatorio para el título de la gráfica}

\item{grupo1}{vector: es el grupo que saldrá del lado derecho del plot}

\item{grupo2}{vector: es el grupo que saldrá del lado izquierdo del plot}

\item{color1}{Color de las barras del grupo 1}

\item{color2}{Color de las barras del grupo 2}
}
\description{
@param bd  Base de datos (ya procesada)
}
