% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_gauge}
\alias{graficar_gauge}
\title{Graficar gauge (donita)}
\usage{
graficar_gauge(
  bd,
  color_principal,
  color_secundario = "gray80",
  escala,
  size_text_pct
)
}
\arguments{
\item{bd}{Base de datos con un único row. La variable a considerar se llama "media"}

\item{color_principal}{Color principal de la barra}

\item{color_secundario}{Color secundario de la barra}

\item{escala}{Máximo y mínimo de los valores que puede tomar "media"}

\item{size_text_pct}{Tamaño del texto dentro del gauge}
}
\value{
Objeto tipo \link[ggplot2:ggplot]{ggplot}
}
\description{
Recibe un \code{\link[tibble:tibble]{tibble()}} y genera un objeto tipo \link[ggplot2:ggplot]{ggplot}. El producto es una grafica de gauge (donita)
que se rellena acuerdo al valor contenido en la escala indicada
}
\examples{
encuestar:::analizar_frecuencias(diseno = encuesta_demo$muestra$diseno, pregunta = "si_voto_24") |> encuestar:::graficar_gauge(color_principal = "red", escala = c(0, 10), size_text_pct = 16)
encuestar:::analizar_frecuencias(diseno = encuesta_demo$muestra$diseno, pregunta = "conoce_pm_astiazaran") |> filter(respuesta == "Sí") |>  encuestar:::graficar_gauge(color_principal = "red", escala = c(0, 1), size_text_pct = 16)
}
