% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_gauge_promedio}
\alias{graficar_gauge_promedio}
\title{Title}
\usage{
graficar_gauge_promedio(
  bd,
  color = "#850D2D",
  escala = c(0, 10),
  size_text_pct
)
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{color}{Color de la barra}

\item{maximo}{Número máximo en la escala}
}
\description{
Title
}
