% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_morena}
\alias{graficar_morena}
\title{Graficar metodología de MORENA con un geom_tile}
\usage{
graficar_morena(atr, personajes, atributos)
}
\arguments{
\item{atr}{Base de datos producto de la función 'analizar_morena'.}

\item{atributos}{Vector de códigos de cortos que identifican los diferentes atributos contenidos en la metodología de MORENA.}

\item{cliente}{Vector de códigos cortos relacionados a los personajes a los cuales se les aplicó la batería de preguntas de MORENA.}
}
\description{
Graficar metodología de MORENA con un geom_tile
}
\examples{
graficar_morena(atr, personajes = c("era", "sasil"), atributos = c("honesto", "opinion"))
graficar_morena(atr, personajes = c("era", "sasil", "jaac"), atributos = atributos)
}
