% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_nube_frecuencias}
\alias{graficar_nube_frecuencias}
\title{Title}
\usage{
graficar_nube_frecuencias(
  bd,
  pregunta,
  n = 100,
  color1 = "#5B0A1C",
  color2 = "#850D2D",
  color3 = "#961B41",
  familia = "Poppins",
  ancho = 600 * 5,
  alto = 600 * 5/1.41
)
}
\arguments{
\item{bd}{Base de datos (ya procesada)}

\item{alto}{}
}
\value{

}
\description{
Title
}
