% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_nubePalabras_hc}
\alias{graficar_nubePalabras_hc}
\title{Graficar nube de palabras clave}
\usage{
graficar_nubePalabras_hc(bd)
}
\arguments{
\item{bd}{Base de datos procesada con la estructura "palabras", "n", y "colores"}
}
\description{
Graficar nube de palabras clave
}
\examples{
graficar_nubePalabras_hc(bd = tibble(palabras = c("hola"), n = c(2), colores = c("#619CFF")))
}
