% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_regionesCategorico}
\alias{graficar_regionesCategorico}
\title{Graficar mapa por categorías}
\usage{
graficar_regionesCategorico(bd, variable)
}
\arguments{
\item{bd}{Base de datos resultado de la función 'calcular_ganadorRegion}

\item{variable}{Variable categórica a graficar}
}
\description{
Graficar mapa por categorías
}
