% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_saldoRegion}
\alias{graficar_saldoRegion}
\title{Graficar saldo de opinión asociado a personajes por región o estrato}
\usage{
graficar_saldoRegion(bd, ordenRegiones, salto_labelRegiones = 5)
}
\arguments{
\item{bd}{Base de datos resultado de la función 'analizar_saldoRegion'}

\item{ordenRegiones}{Vector que indica el orden de las columnas del geom_tile}

\item{salto_labelRegiones}{Parámetro 'width' usado por stringr::str_wrap en las etiquetas superiores}
}
\description{
Graficar saldo de opinión asociado a personajes por región o estrato
}
