% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Graficas}
\alias{Graficas}
\title{Esta es la clase de Pregunta2}
\description{
Esta es la clase de Pregunta2

Esta es la clase de Pregunta2
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Graficas-new}{\code{Graficas$new()}}
\item \href{#method-Graficas-clone}{\code{Graficas$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-new"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$new(encuesta, tema = tema_default)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Graficas-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Graficas-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Graficas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
