% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clase_resultados.R
\name{Modelo}
\alias{Modelo}
\title{Esta es la clase de Modelo}
\description{
Esta es la clase de Modelo
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Modelo-initialize}{\code{Modelo$new()}}
    \item \href{#method-Modelo-correspondencia}{\code{Modelo$correspondencia()}}
    \item \href{#method-Modelo-componentesPrincipales}{\code{Modelo$componentesPrincipales()}}
    \item \href{#method-Modelo-clone}{\code{Modelo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-initialize}{}}}
\subsection{\code{Modelo$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Modelo$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-correspondencia"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-correspondencia}{}}}
\subsection{\code{Modelo$correspondencia()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Modelo$correspondencia(
  var1,
  var2,
  legenda1 = NULL,
  legenda2 = NULL,
  colores = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-componentesPrincipales"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-componentesPrincipales}{}}}
\subsection{\code{Modelo$componentesPrincipales()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Modelo$componentesPrincipales(variables)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-clone}{}}}
\subsection{\code{Modelo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Modelo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
