% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{obtener_ubicacionEfectiva_surveyToGo}
\alias{obtener_ubicacionEfectiva_surveyToGo}
\title{Obtener telemetría de intento efectivo de encuesta}
\usage{
obtener_ubicacionEfectiva_surveyToGo(bd_respuestas, id, intento_efectivo)
}
\arguments{
\item{bd_respuestas}{Base de datos con formato exportado de SurveyToGo}

\item{id}{Id único que identifica cada entrevista, generalmente se llama SbjNum}

\item{intento_efectivo}{Veces que se intenta levantar una entrevista hasta que esta comienza}
}
\description{
Obtener telemetría de intento efectivo de encuesta
}
