% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{pclave_combinaciones_saldo}
\alias{pclave_combinaciones_saldo}
\title{Title}
\usage{
pclave_combinaciones_saldo(bd_texto, tipo, n_palabras)
}
\arguments{
\item{n_palabras}{}
}
\value{

}
\description{
Title
}
