% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Pregunta2}
\alias{Pregunta2}
\title{Esta es la clase de Pregunta2}
\description{
Esta es la clase de Pregunta2

Esta es la clase de Pregunta2
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Pregunta2-new}{\code{Pregunta2$new()}}
\item \href{#method-Pregunta2-clone}{\code{Pregunta2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta2-new"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta2-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta2$new(encuesta, tema = tema_default)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
