% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_cuestionario.R
\name{resumen_cuestionario}
\alias{resumen_cuestionario}
\title{Title}
\usage{
resumen_cuestionario(dicc)
}
\value{
\item{a}{Grafica de cascada para saber cuantas preguntas por bloque hay y en total.}
\item{b}{Grafica indicando cuales son las etiquetas y cuantos aspectos hay para cada pregunta de aspectos.}
\item{c}{Grafica gant, se muestra todo el cuestionario con sus respectivas categorias en una sola grafica.}
}
\description{
Title
}
