% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analisis_correspondencia}
\alias{analisis_correspondencia}
\title{Title}
\usage{
analisis_correspondencia(
  var1,
  var2,
  legenda1 = NULL,
  legenda2 = NULL,
  diseno,
  colores = NULL
)
}
\arguments{
\item{colores}{}
}
\description{
Title
}
