% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_candidatoPartido}
\alias{analizar_candidatoPartido}
\title{Analizar partido asociado a un candidato}
\usage{
analizar_candidatoPartido(
  diseno,
  diccionario,
  llave_partido,
  llave_conocimiento,
  respuesta_conoce,
  candidatos,
  corte_otro
)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{diccionario}{Cuestionario de la encuesta en formato de procesamiento requerido}

\item{llave_partido}{Patrón que comparten las variables asociadas a las preguntas que relacionan un personaje con un partido político}

\item{llave_conocimiento}{Patrón que comparten las variables relacionadas a preguntas sobre el conocimiento de un personaje}

\item{respuesta_conoce}{Filtro sobre el cuál se evalúa la asociación de un personaje a un partído político}

\item{candidatos}{Vector de nombres cortos asociados a uno o más personajes sobre los cuáles se preguntó su asociación partidista}

\item{corte_otro}{Parámetro 'prop' de la función 'fct_lump' de la paquetería 'forcats' usado para agrupar valores pequeños de partidos políticos}
}
\description{
Analizar partido asociado a un candidato
}
\examples{
analizar_candidatoPartido(diseno = self\%diseno, llave_partido = "partido", llave_conocimiento = "conocimiento", respuesta_conoce = "Sí", candidatos = c("era", "sasil"))
}
