% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_cruce}
\alias{analizar_cruce}
\title{Analizar cruce entre dos variables}
\usage{
analizar_cruce(diseno, variable_principal, variable_secundaria, vartype)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{variable_secundaria}{}
}
\description{
Analizar cruce entre dos variables
}
