% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_cruce_aspectos}
\alias{analizar_cruce_aspectos}
\title{Analizar cruce por puntos}
\usage{
analizar_cruce_aspectos(
  diseno,
  variable_principal,
  variables_secundarias,
  filtro_variables_secundarias,
  vartype
)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{vartype}{}
}
\description{
Analizar cruce por puntos
}
