% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_cruce_aspectos}
\alias{analizar_cruce_aspectos}
\title{Analizar cruce entre multiples variables cuyos nombres compartan valores unicos
de un diseno muestral construido con la paqueteria \code{survey}}
\usage{
analizar_cruce_aspectos(
  diseno,
  variable_principal,
  variables_secundarias,
  filtro_variables_secundarias,
  vartype
)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{variable_principal}{Variable principal que agrupa los resultados de las estimaciones}

\item{variables_secundarias}{Variables secundarias que compartan valores unicos}

\item{filtro_variables_secundarias}{Valor unico de las variables secundarias de interes a mostrar en resultados}

\item{vartype}{Parametro de \code{\link[srvyr:survey_mean]{srvyr::survey_mean()}} que reporta la variabilidad de la estimacion}
}
\value{
\code{\link[=tibble]{tibble()}} con las estimaciones de frecuencia para cada valor unico entre las variables
}
\description{
Analizar cruce entre multiples variables cuyos nombres compartan valores unicos
de un diseno muestral construido con la paqueteria \code{survey}
}
\examples{
analizar_cruce_aspectos(diseno = encuesta_demo$muestra$diseno, variable_principal = "sexo", variables_secundarias = c("conoce_pm_lia", "conoce_pm_javier"), filtro_variables_secundarias = "Sí", vartype = "cv")
}
