% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias}
\alias{analizar_frecuencias}
\title{Calcula la media de un diseno muestral construiido con la libreria survvey}
\usage{
analizar_frecuencias(diseno, pregunta)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas}

\item{pregunta}{Nombre de la variable a calcular la estimación de proporciones de valores en la base de datos}
}
\value{
\item{estimacion}{Tabla con las estimaciones de frecuencia para cada categoría respondida}
}
\description{
Calcula la media de un diseno muestral construiido con la libreria survvey
}
\examples{
analizar_frecuencias(diseno = as_survey_design(encuesta_demo$muestra$diseno), pregunta = "sexo")
}
