% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias}
\alias{analizar_frecuencias}
\title{Analizar frecuencias}
\usage{
analizar_frecuencias(diseno, pregunta)
}
\arguments{
\item{diseno}{Diseño muestral que contiene la información de pesos de cada individuo}

\item{pregunta}{Nombre de la variable asociada al análisis de la muestra}
}
\value{
\item{estimacion}{Tabla con las estimaciones de frecuencia para cada categoría respondida}
}
\description{
Analizar frecuencias
}
