% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias}
\alias{analizar_frecuencias}
\title{Title}
\usage{
analizar_frecuencias(encuesta, pregunta)
}
\arguments{
\item{pregunta}{}
}
\value{
\item{estimacion}{Tabla con las estimaciones de frecuencia para cada categoría respondida}
}
\description{
Title
}
