% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias}
\alias{analizar_frecuencias}
\title{Analizar frecuencias}
\usage{
analizar_frecuencias(diseno, pregunta)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas}

\item{pregunta}{Nombre de la variable a calcular la estimación de proporciones de valores en la base de datos}
}
\value{
\item{estimacion}{Tabla con las estimaciones de frecuencia para cada categoría respondida}
}
\description{
Analizar frecuencias
}
\examples{
analizar_frecuencias(diseno = as_survey_design(diseno), pregunta = sexo)
analizar_frecuencias(diseno = as_survey_design(encuesta$muestra$diseno), pregunta = sexo)
}
