% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias_aspectos}
\alias{analizar_frecuencias_aspectos}
\title{Title}
\usage{
analizar_frecuencias_aspectos(encuesta, pregunta, aspectos)
}
\arguments{
\item{aspectos}{}
}
\description{
Title
}
