% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_frecuencias_multirespuesta}
\alias{analizar_frecuencias_multirespuesta}
\title{Analizar frecuencias a multiples variables asociadas a una misma pregunta con opcion a multirespuesta
de un diseno muestral construido con la paqueteria \code{survey}}
\usage{
analizar_frecuencias_multirespuesta(diseno, patron_inicial)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas}

\item{patron_inicial}{Cadena de texto en comun (inicial) entre los nombres de las variables asociadas a la pregunta}
}
\value{
\code{\link[=tibble]{tibble()}} con las estimaciones de frecuencia para cada valor unico entre las variables
}
\description{
Analizar frecuencias a multiples variables asociadas a una misma pregunta con opcion a multirespuesta
de un diseno muestral construido con la paqueteria \code{survey}
}
\examples{
analizar_frecuencia_multirespuesta(diseno = encuesta_demo$muestra$diseno, patron_inicial = "medios_com_O")
}
