% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{analizar_sankey}
\alias{analizar_sankey}
\title{Analizar sankey}
\usage{
analizar_sankey(diseno, variables, filtro_var1, filtro_var2)
}
\arguments{
\item{diseno}{Diseno muestral que contiene los pesos por individuo y las variables relacionadas.}

\item{variables}{Vector que contiene las llaves de las cuales se va a hacer el cruce}
}
\description{
Analizar sankey
}
\examples{
analizar_sankey(diseno = as_survey_design(diseno), var_1 = sexo, var_2 = conocimento_sheinbaum)
analizar_frecuencias(diseno = as_survey_design(encuesta$muestra$diseno), var_1 = sexo, var_2 = conocimento_sheinbaum)
}
