% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_mediaMovil}
\alias{calcular_mediaMovil}
\title{Calcular media movil de una variable}
\usage{
calcular_mediaMovil(bd_resultados, variable, sin_peso = T)
}
\arguments{
\item{sin_peso}{}
}
\description{
Calcular media movil de una variable
}
