% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_mediaMovil_region}
\alias{calcular_mediaMovil_region}
\title{Calcular media movil de una variable por region}
\usage{
calcular_mediaMovil_region(
  bd_resultados,
  variable,
  sin_peso,
  valores_interes,
  variable_region
)
}
\arguments{
\item{variable_region}{}
}
\description{
Calcular media movil de una variable por region
}
