% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_tabla_candidatoOpinion}
\alias{calcular_tabla_candidatoOpinion}
\title{Title}
\usage{
calcular_tabla_candidatoOpinion(
  diseno,
  diccionario,
  patron_opinion,
  patron_conocimiento,
  aspectos,
  filtro_conocimiento,
  orden_opinion,
  ns_nc,
  salto_respuestas
)
}
\arguments{
\item{ns_nc}{}
}
\description{
Title
}
