% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_surveytogo.R
\name{calcular_tabla_votoCruzado}
\alias{calcular_tabla_votoCruzado}
\title{Title}
\usage{
calcular_tabla_votoCruzado(diseno, var1, var2, filtro_var2)
}
\arguments{
\item{var2}{}
}
\description{
Title
}
