% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clase_resultados.R
\name{Cruce}
\alias{Cruce}
\title{Esta es la clase de Cruce}
\description{
Esta es la clase de Cruce

Esta es la clase de Cruce
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cruce-new}{\code{Cruce$new()}}
\item \href{#method-Cruce-bloques}{\code{Cruce$bloques()}}
\item \href{#method-Cruce-lineas}{\code{Cruce$lineas()}}
\item \href{#method-Cruce-lolipop_diferencias}{\code{Cruce$lolipop_diferencias()}}
\item \href{#method-Cruce-sankey_categorica}{\code{Cruce$sankey_categorica()}}
\item \href{#method-Cruce-tabla_votoCruzado}{\code{Cruce$tabla_votoCruzado()}}
\item \href{#method-Cruce-clone}{\code{Cruce$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-bloques"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-bloques}{}}}
\subsection{Method \code{bloques()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$bloques(
  variable_principal,
  variable_secundaria,
  colores_variable_secundaria,
  filter = NULL,
  vartype = "cv",
  linea_grosor = 2,
  linea_color = "white",
  na_rm = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-lineas"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-lineas}{}}}
\subsection{Method \code{lineas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$lineas(
  variable_principal,
  variable_secundaria,
  orden_variable_principal,
  valores_variable_secundaria = NULL,
  colores_variable_secundaria,
  limits = c(0, 0.75),
  wrap_x = 25,
  wrap_legend = 25,
  text_nudge_y = 0.01,
  size_text = 8,
  size_text_x = 16,
  size_text_y = 14,
  size_text_legend = 14,
  na_rm = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-lolipop_diferencias"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-lolipop_diferencias}{}}}
\subsection{Method \code{lolipop_diferencias()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$lolipop_diferencias(
  variable_principal,
  variables_secundarias,
  filtro_variables_secundarias,
  orden_variablePrincipal,
  colores_variables_secundarias,
  caption,
  nudge_x = 0.05,
  size_geom_text = 6,
  invertir_variables = F,
  vartype = "cv",
  limits = c(0, 0.75),
  wrap_y = 25,
  wrap_caption = 25,
  size_text_x = 16,
  size_text_y = 16,
  size_text_caption = 16,
  size_text_legend = 16,
  ver_diferencias = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-sankey_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-sankey_categorica}{}}}
\subsection{Method \code{sankey_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$sankey_categorica(
  variables = NULL,
  colores,
  size_text_cat = 6,
  width_text = 15,
  omitir_valores_variable1 = NULL,
  omitir_valores_variable2 = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-tabla_votoCruzado"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-tabla_votoCruzado}{}}}
\subsection{Method \code{tabla_votoCruzado()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$tabla_votoCruzado(
  var1,
  var2,
  filtro_var2 = NULL,
  etiquetas = c("variable 1", "variable 2"),
  colores_var1,
  colores_var2,
  size_text_header = 18,
  size_text_body = 14,
  salto = 20,
  na_rm = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
