% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Cruce}
\alias{Cruce}
\title{Esta es la clase de Cruce}
\description{
Esta es la clase de Cruce

Esta es la clase de Cruce
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cruce-new}{\code{Cruce$new()}}
\item \href{#method-Cruce-sankey_categorica}{\code{Cruce$sankey_categorica()}}
\item \href{#method-Cruce-puntos}{\code{Cruce$puntos()}}
\item \href{#method-Cruce-puntosMultiples}{\code{Cruce$puntosMultiples()}}
\item \href{#method-Cruce-brechasDuales}{\code{Cruce$brechasDuales()}}
\item \href{#method-Cruce-brechasMultiples}{\code{Cruce$brechasMultiples()}}
\item \href{#method-Cruce-bloques}{\code{Cruce$bloques()}}
\item \href{#method-Cruce-clone}{\code{Cruce$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-sankey_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-sankey_categorica}{}}}
\subsection{Method \code{sankey_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$sankey_categorica(
  variables = NULL,
  colores,
  size_text_cat = 6,
  omitir_valores_variable1 = NULL,
  omitir_valores_variable2 = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-puntos"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-puntos}{}}}
\subsection{Method \code{puntos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$puntos(cruce, variables, vartype = "se", valor_variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-puntosMultiples"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-puntosMultiples}{}}}
\subsection{Method \code{puntosMultiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$puntosMultiples(
  variablePrincipal,
  variablesSecundarias,
  valor_variblesSecundarias,
  orden_variablePrincipal,
  invertirVariables = F,
  vartype = "cv"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-brechasDuales"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-brechasDuales}{}}}
\subsection{Method \code{brechasDuales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$brechasDuales(
  var1,
  var2_filtro,
  filtro,
  vartype = "cv",
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-brechasMultiples"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-brechasMultiples}{}}}
\subsection{Method \code{brechasMultiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$brechasMultiples(
  por_grupo,
  variables,
  vartype = "cv",
  valor_variables,
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-bloques"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-bloques}{}}}
\subsection{Method \code{bloques()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$bloques(
  cruce,
  variable,
  vartype = "cv",
  filter = NULL,
  linea_grosor = 2,
  linea_color = "white"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cruce-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cruce-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cruce$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
