% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Cuestionario}
\alias{Cuestionario}
\title{Esta es la clase cuestionario}
\description{
Esta es la clase cuestionario

Esta es la clase cuestionario
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cuestionario$new()}}
\item \href{#method-aprobar}{\code{Cuestionario$aprobar()}}
\item \href{#method-checar_pregunta}{\code{Cuestionario$checar_pregunta()}}
\item \href{#method-resumen}{\code{Cuestionario$resumen()}}
\item \href{#method-clone}{\code{Cuestionario$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cuestionario$new(documento)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aprobar"></a>}}
\if{latex}{\out{\hypertarget{method-aprobar}{}}}
\subsection{Method \code{aprobar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cuestionario$aprobar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checar_pregunta"></a>}}
\if{latex}{\out{\hypertarget{method-checar_pregunta}{}}}
\subsection{Method \code{checar_pregunta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cuestionario$checar_pregunta(pregunta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resumen"></a>}}
\if{latex}{\out{\hypertarget{method-resumen}{}}}
\subsection{Method \code{resumen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cuestionario$resumen()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cuestionario$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
