% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Descriptiva}
\alias{Descriptiva}
\title{Esta es la clase de Descriptiva}
\description{
Esta es la clase de Descriptiva

Esta es la clase de Descriptiva
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Descriptiva-new}{\code{Descriptiva$new()}}
\item \href{#method-Descriptiva-barras_categorica}{\code{Descriptiva$barras_categorica()}}
\item \href{#method-Descriptiva-barras_aspectos}{\code{Descriptiva$barras_aspectos()}}
\item \href{#method-Descriptiva-barras_multirespuesta}{\code{Descriptiva$barras_multirespuesta()}}
\item \href{#method-Descriptiva-gauge_numerica}{\code{Descriptiva$gauge_numerica()}}
\item \href{#method-Descriptiva-gauge_categorica}{\code{Descriptiva$gauge_categorica()}}
\item \href{#method-Descriptiva-intervalo_numerica}{\code{Descriptiva$intervalo_numerica()}}
\item \href{#method-Descriptiva-lollipops_categorica}{\code{Descriptiva$lollipops_categorica()}}
\item \href{#method-Descriptiva-lollipops_multirespuesta}{\code{Descriptiva$lollipops_multirespuesta()}}
\item \href{#method-Descriptiva-clone}{\code{Descriptiva$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-new"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-barras_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-barras_categorica}{}}}
\subsection{Method \code{barras_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$barras_categorica(
  codigo,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0,
  pct_otros = 0.01,
  orden_respuestas = NA
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-barras_aspectos"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-barras_aspectos}{}}}
\subsection{Method \code{barras_aspectos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$barras_aspectos(
  patron_inicial,
  aspectos = NULL,
  salto = 20,
  filtro = "respuesta == 'Sí'",
  porcentajes_fuera = F,
  desplazar_porcentajes = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-barras_multirespuesta"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-barras_multirespuesta}{}}}
\subsection{Method \code{barras_multirespuesta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$barras_multirespuesta(
  patron_inicial,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-gauge_numerica"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-gauge_numerica}{}}}
\subsection{Method \code{gauge_numerica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$gauge_numerica(
  codigo,
  color = "#850D2D",
  escala = c(0, 10),
  size_text_pct = 14
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-gauge_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-gauge_categorica}{}}}
\subsection{Method \code{gauge_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$gauge_categorica(
  codigo,
  filtro,
  color = "#850D2D",
  escala = c(0, 1),
  size_text_pct = 14
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-intervalo_numerica"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-intervalo_numerica}{}}}
\subsection{Method \code{intervalo_numerica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$intervalo_numerica(
  patron,
  aspectos,
  escala = c(0, 10),
  point_size = 1,
  text_point_size = 8
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-lollipops_categorica"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-lollipops_categorica}{}}}
\subsection{Method \code{lollipops_categorica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$lollipops_categorica(
  codigo,
  orden = NULL,
  limite_graf = 1,
  width_cats = 15,
  size = 3,
  size_pct = 6,
  pct_otros = 0.01
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-lollipops_multirespuesta"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-lollipops_multirespuesta}{}}}
\subsection{Method \code{lollipops_multirespuesta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$lollipops_multirespuesta(
  patron_inicial,
  orden = NULL,
  limite_graf = 1,
  width_cats = 15,
  size = 3,
  size_pct = 6
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Descriptiva-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Descriptiva-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Descriptiva$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
