% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Encuesta}
\alias{Encuesta}
\title{Esta es la clase Encuesta}
\description{
La clase Encuesta contiene todos los métodos, y atributos que puede tener una encuesta.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{respuestas}}{Base de datos de respuestas}

\item{\code{muestra}}{Base de datos de muestra.}

\item{\code{dicionario}}{Base de datos de diccionario}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dicionario}}{Base de datos de diccionario}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Encuesta$new()}}
\item \href{#method-clone}{\code{Encuesta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$new(
  respuestas = NA,
  muestra = NA,
  auditoria_telefonica = NA,
  cuestionario = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{respuestas}}{Name of the person}

\item{\code{diccionario}}{Hair colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
