% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Encuesta}
\alias{Encuesta}
\title{Esta es la clase Encuesta}
\description{
La clase Encuesta contiene todos los metodos, y atributos que puede tener una encuesta.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{respuestas}}{Base de datos de respuestas}

\item{\code{muestra}}{Base de datos de muestra.}

\item{\code{dicionario}}{Base de datos de diccionario}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dicionario}}{Base de datos de diccionario}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Encuesta-new}{\code{Encuesta$new()}}
\item \href{#method-Encuesta-simular_surveytogo}{\code{Encuesta$simular_surveytogo()}}
\item \href{#method-Encuesta-error_muestral_maximo}{\code{Encuesta$error_muestral_maximo()}}
\item \href{#method-Encuesta-exportar_entregable}{\code{Encuesta$exportar_entregable()}}
\item \href{#method-Encuesta-clone}{\code{Encuesta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-new}{}}}
\subsection{Method \code{new()}}{
Create a person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$new(
  respuestas = NA,
  n_simulaciones = NULL,
  opinometro_id = NULL,
  pool = NULL,
  bd_categorias = NULL,
  quitar_vars = NA,
  muestra = NA,
  auditoria_telefonica = NA,
  bd_correcciones = NULL,
  cuestionario = NA,
  shp = NA,
  tipo_encuesta = "ine",
  mantener = NA,
  patron = NA,
  auditar = NA,
  vars_tendencias = NULL,
  sin_peso = F,
  rake = T,
  mantener_falta_coordenadas = F,
  dir_app = "auditoria"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{respuestas}}{Name of the person}

\item{\code{diccionario}}{Hair colour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-simular_surveytogo"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-simular_surveytogo}{}}}
\subsection{Method \code{simular_surveytogo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$simular_surveytogo(cuestionario, n, diseño, shp)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-error_muestral_maximo"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-error_muestral_maximo}{}}}
\subsection{Method \code{error_muestral_maximo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$error_muestral_maximo(quitar_patron = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-exportar_entregable"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-exportar_entregable}{}}}
\subsection{Method \code{exportar_entregable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$exportar_entregable(
  carpeta = "entregables",
  agregar = NULL,
  quitar = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Encuesta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Encuesta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Encuesta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
