% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estandarizar_texto}
\alias{estandarizar_texto}
\title{Estandarizar texto para imitar resultados de la categorizacion con el bot}
\usage{
estandarizar_texto(
  texto,
  quitar_acentos = TRUE,
  quitar_caracteres_especiales = TRUE,
  minusculas = TRUE
)
}
\arguments{
\item{minusculas}{}
}
\description{
Estandarizar texto para imitar resultados de la categorizacion con el bot
}
