% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generarGlosario_preguntaAbierta}
\alias{generarGlosario_preguntaAbierta}
\title{Title}
\usage{
generarGlosario_preguntaAbierta(
  folder = "./data/",
  prefijo = "glosario_",
  bd_categorias_raw,
  variable
)
}
\arguments{
\item{variable}{}
}
\description{
Title
}
