% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras}
\alias{graficar_barras}
\title{Gráfica de barras horizontales ordenadas}
\usage{
graficar_barras(
  bd,
  salto = 20,
  porcentajes_fuera = F,
  desplazar_porcentajes = 0,
  orden_respuestas = NA
)
}
\arguments{
\item{bd}{Base de datos con una variable categórica (respuesta) y una numérica (media).}

\item{salto}{Número entero, se aplica un stringr::str_wrap a la variable categórica.}

\item{porcentajes_fuera}{Si es T, las labels de los porcentajes aparecen fuera (o sobre) las barras.}

\item{desplazar_porcentajes}{Si porcentajes_fuera es T, este parametro ajusta las etiquetas de texto.}
}
\description{
Gráfica de barras horizontales ordenadas
}
\examples{
graficar_barras(bd, salto = 13)
graficar_barras(bd, porcentajes_fuera = T, desplazar_porcentajes = 0.1)
}
