% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_frecuencia}
\alias{graficar_barras_frecuencia}
\title{Title}
\usage{
graficar_barras_frecuencia(bd, titulo, fill = NULL, nota = "", colores)
}
\arguments{
\item{bd}{debe provenir de la función de analizar_frecuencias}

\item{titulo}{Es un parámetro obligatorio para}

\item{nota}{Si de desea añadir una nota al pie del plot}

\item{colores}{Vector de colores del plot}

\item{tamano_letra}{Tamaño de letra de los elementos del plot}

\item{tamano_titulo}{Tamaño de letra solamente del título}

\item{color_titulo}{Color de la letra del título}

\item{familia}{Familia tipográfica de los elementos del plot}

\item{color_etiqueta}{Color de la letra de las etiquetas de datos}
}
\value{

}
\description{
Title
}
