% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_barras_saldo}
\alias{graficar_barras_saldo}
\title{Title}
\usage{
graficar_barras_saldo(
  bd,
  orden,
  grupo_positivo,
  grupo_negativo,
  Regular,
  colores,
  salto_respuestas,
  salto_tema,
  caption_opinion,
  size_text_cat,
  size_pct,
  size_caption_opinion,
  tema = encuestar:::tema_morant()
)
}
\arguments{
\item{tema}{}
}
\description{
Title
}
