% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_candidato_opinion}
\alias{graficar_candidato_opinion}
\title{Graficar candidato opinión}
\usage{
graficar_candidato_opinion(
  bd,
  ns_nc,
  regular,
  grupo_positivo,
  grupo_negativo,
  colores,
  color_nsnc,
  burbuja,
  color_burbuja,
  caption_opinion,
  caption_nsnc,
  caption_burbuja,
  size_caption_opinion,
  size_caption_nsnc,
  size_caption_burbuja,
  size_text_cat,
  size_pct,
  orden_resp,
  salto = 200,
  tema,
  mostrar_nsnc = T
)
}
\arguments{
\item{bd}{Base de datos con estructura producida por analizar_frecuencias_aspectos.}

\item{ns_nc}{Valor de la variable 'respuesta' asociado a "No sabe" o "No contesta" en la pregunta relativa.}

\item{regular}{Valor de la variable 'respuesta' asociado a "Regular" en la pregunta relativa.}

\item{grupo_positivo}{Conjunto de valores de la  variable 'respuesta' considerados como positivos.}

\item{grupo_negativo}{Conjunto de valores de la  variable 'respuesta' considerados como negativos}

\item{colores}{Vector ordenado de colores asociados al grupo negativo, regular y positivo.}

\item{burbuja}{Base de datos con estructura producida por analizar_frecuencias_aspectos filtrada por un valor de interés. Disponible en el enviroment.}

\item{color_burbuja}{Color de los puntos asociados a la base de datos 'burbuja'}

\item{caption_opinion}{Caption de la gráfica que visualiza los datos de bd}

\item{caption_nsnc}{Caption de la gráfica que visualiza los datos de bd separando el valor asociado a "No sabe" o "No contesta" en la pregunta relativa.}

\item{caption_burbuja}{Caption de la gráfica que visualiza los datos de burbuja}

\item{size_caption_opinion}{Tamaño del texto del caption_opinion}

\item{size_caption_nsnc}{Tamaño del texto del caption_nsnc}

\item{size_caption_burbuja}{Tamaño del texto del caption_burbuja}

\item{size_text_cat}{Tamaño del texto de las categorías diferentes categorías de la variable 'tema' de la base de datos 'bd'}

\item{orden_resp}{Vector ordenado de los posibles valores de la variable 'respuesta'}

\item{salto}{Parámetro usado por la función str_wrap de la paquetería stringr aplicado a la variable 'tema' de la base 'bd'}

\item{tema}{Tema de la gráfica asociado a la paquetería 'encuestar'}
}
\description{
Graficar candidato opinión
}
\examples{
graficar_candidato_opinion(bd, ns_nc = "Ns/Nc", regular = "Regular", grupo_positivo = "Buena", grupo_negativo = "Mala", colores = c("red", "yellow", "green", "gray70"), burbuja = burbuja, color_burbuja = "blue", caption_opinion = "", caption_nsnc = "Ns/Nc", caption_burbuja = "Nivel de conocimiento", size_caption_opinion = 12, size_caption_nsnc = 12, size_caption_burbuja = 12, size_caption_cat = 12, orden_resp = c("Mala", "Regular", "Buena"), tema = self$tema)
}
