% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_candidatoPartido}
\alias{graficar_candidatoPartido}
\title{Graficar el partido político con el que asocian a uno o varios personajes}
\usage{
graficar_candidatoPartido(
  bases,
  cliente,
  tipo_conoce,
  colores_candidato,
  solo_respondidos = T,
  colores_partido,
  tema,
  corte_vis = 0,
  size_text = 6,
  size_text_conocimiento = 6
)
}
\arguments{
\item{bases}{Lista. 'bases$conoce' contiene la estimación del conocimiento sobre algún personaje. 'bases$partido' tiene la información sobre la asociación a un partido político por personaje}

\item{cliente}{Vector de códigos cortos que asocian a un personaje con una variable. Por ejemplo, 'personajeC' resaltará entre el conjunto 'personajeA', 'personajeB', 'personajeC'.}

\item{tipo_conoce}{Tipo de gráfica a mostrar en la estimación de conocimiento. De forma predeterminada son barras, el otro valor son 'intervalos'}

\item{colores_candidato}{Vector que asigna un color a un candidato de acuerdo al nombre largo (tema)}

\item{solo_respondidos}{Logical. Omite los partidos en los cuales el personaje no tiene ninguna asociación}

\item{colores_partido}{Vector que asigna un color a cada partido de acuerdo al nombre largo (tema)}

\item{tema}{Tema de la gráfica asociado a la paquetería 'encuestar'}

\item{size_text_conocimiento}{Valor de tipo entero. Tamaño de los porcentajes en la grafica de conocimiento.}
}
\description{
Recibe un \code{\link[tibble:tibble]{tibble()}} y genera un objeto tipo \link[ggplot2:ggplot]{ggplot}. El producto es una grafica tipo barras horizontales
combinada con una grafica tipo intervalos. La grafica tipo barras muestra los porcentajes con de
con los que asocian a determinado personaje. La grafica de intervalos muestra el nivel de conocimiento
del mismo.
}
