% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_candidatoPartido}
\alias{graficar_candidatoPartido}
\title{Graficar el partido político con el que asocian a uno o varios personajes}
\usage{
graficar_candidatoPartido(
  bases,
  cliente,
  tipo_conoce,
  colores_candidato,
  solo_respondidos = T,
  colores_partido,
  tema
)
}
\arguments{
\item{bases}{Lista. 'bases$conoce' contiene la estimación del conocimiento sobre algún personaje. 'bases$partido' tiene la información sobre la asociación a un partido político por personaje}

\item{cliente}{Vector de códigos cortos que asocian a un personaje con una variable. Por ejemplo, 'personajeC' resaltará entre el conjunto 'personajeA', 'personajeB', 'personajeC'.}

\item{tipo_conoce}{Tipo de gráfica a mostrar en la estimación de conocimiento. De forma predeterminada son barras, el otro valor son 'intervalos'}

\item{colores_candidato}{Vector que asigna un color a un candidato de acuerdo al nombre largo (tema)}

\item{solo_respondidos}{Logical. Omite los partidos en los cuales el personaje no tiene ninguna asociación}

\item{colores_partido}{Vector que asigna un color a cada partido de acuerdo al nombre largo (tema)}

\item{tema}{Tema de la gráfica asociado a la paquetería 'encuestar'}
}
\description{
Graficar el partido político con el que asocian a uno o varios personajes
}
\examples{
graficar_candidato_partido(bases, clientes = c("era", "sasil"), tipo_conoce = "intervalos", colores_candidato = colores_candidato, colores_partido = colores_partido, tema = self$tema)
}
