% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_candidatoSaldo}
\alias{graficar_candidatoSaldo}
\title{Graficar saldo de opinión por personaje}
\usage{
graficar_candidatoSaldo(
  bd,
  grupo_positivo = c("Buena", "Muy buena"),
  grupo_negativo = c("Mala", "Muy mala"),
  color_positivo = "green",
  color_negativo = "red"
)
}
\arguments{
\item{bd}{Base de datos resultado de la función 'calcular_saldoOpinion'.}

\item{grupo_positivo}{Valores posibles dentro de la variable 'grupo' que los identifica de manera arbitraria como positivos.}

\item{grupo_negativo}{Valores posibles dentro de la variable 'grupo' que los identifica de manera arbitraria como negativos.}

\item{color_positivo}{Color asociado al grupo positivo.}

\item{color_negativo}{Color asociado al grupo negativo.}
}
\value{
Objeto tipo \link{ggplot}
}
\description{
Graficar saldo de opinión por personaje
}
