% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_bloques}
\alias{graficar_cruce_bloques}
\title{Title}
\usage{
graficar_cruce_bloques(
  bd,
  cruce,
  variable,
  colores_variable_secundaria,
  vartype,
  filter,
  linea_grosor,
  linea_color
)
}
\arguments{
\item{linea_color}{}
}
\description{
Title
}
