% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_cruce_bloques}
\alias{graficar_cruce_bloques}
\title{Graficar cruce de una variable vs otra con opción a filtro}
\usage{
graficar_cruce_bloques(bd, cruce, variable, vartype, filter)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_cruceBrechas'}

\item{cruce}{Variable principal por la cual se hace el cruce}

\item{variable}{Variable secundaria para hacer análisis con la primaria}

\item{filter}{Filtro aplicable al parámetro 'cruce'}
}
\description{
Graficar cruce de una variable vs otra con opción a filtro
}
