% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_gauge}
\alias{graficar_gauge}
\title{Graficar gauge}
\usage{
graficar_gauge(
  bd,
  color_principal,
  color_secundario = "gray80",
  escala,
  size_text_pct
)
}
\arguments{
\item{bd}{Base de datos con un único row. La variable a considerar se llama "media"}

\item{color_principal}{Color principal de la barra}

\item{color_secundario}{Color secundario de la barra}

\item{escala}{Máximo y mínimo de los valores que puede tomar "media"}

\item{size_text_pct}{Tamaño del texto dentro del gauge}
}
\description{
Graficar gauge
}
\examples{
graficar_gauge_promedio(bd = bd_procesada, color_principal = "red", escala = c(0, 10), size_text_pct = 8)
graficar_gauge_promedio(bd = bd_procesada, color_principal = "pink", color_secundario = "brown", escala = c(1, 7), size_text_pct = 14)
}
