% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_heatmap}
\alias{graficar_heatmap}
\title{Title}
\usage{
graficar_heatmap(
  bd,
  orden_x,
  orden_y,
  color = "blue",
  caption,
  salto_x,
  salto_y,
  size_text_x = 14,
  size_text_y = 16,
  size_text_caption = 14
)
}
\arguments{
\item{size_text_caption}{}
}
\description{
Title
}
