% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_intervalo_numerica}
\alias{graficar_intervalo_numerica}
\title{Graficar intervalos numérica}
\usage{
graficar_intervalo_numerica(
  bd,
  escala = c(0, 1),
  point_size = 1,
  text_point_size = 8
)
}
\arguments{
\item{bd}{Base de datos con una variable categórica (respuesta) y una numérica (media).}

\item{point_size}{Tamaño del punto que indica el promedio de la estimación.}

\item{text_point_size}{Tamaño del texto que acompaña el valor de la estimación}
}
\description{
Graficar intervalos numérica
}
\examples{
graficar_intervalo_numerica(bd = bd, point_size = 2, text_point_size = 14)
}
