% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_lineas}
\alias{graficar_lineas}
\title{Title}
\usage{
graficar_lineas(
  bd,
  orden_var_x,
  colores_var_y,
  salto_x,
  salto_legend,
  limits = c(0, 0.75),
  text_nudge_y = 0.01,
  size_text = 8
)
}
\arguments{
\item{size_text_legend}{}
}
\description{
Title
}
