% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_lolipop_diferencias}
\alias{graficar_lolipop_diferencias}
\title{Graficar cruce de una variable vs varias variables pero filtradas}
\usage{
graficar_lolipop_diferencias(
  bd,
  orden_variablePrincipal,
  colores_variables_secundarias,
  nudge_x = 0.05,
  size_geom_text = 6,
  caption,
  wrap_y,
  wrap_caption,
  limits
)
}
\arguments{
\item{bd}{Base de datos producto de la función 'analizar_crucePuntos'}

\item{orden_variablePrincipal}{Factor de la variable principal}
}
\description{
Graficar cruce de una variable vs varias variables pero filtradas
}
