% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_lollipops}
\alias{graficar_lollipops}
\title{Graficar lollipop}
\usage{
graficar_lollipops(
  bd,
  orden = NULL,
  limite_graf = 1,
  width_cats = 15,
  size = 3,
  size_pct = 6
)
}
\arguments{
\item{bd}{Base de datos procesada con la función analizar_sankey}

\item{orden}{Default orden = NULL  Recibe un vector como parametro para indicar el orden de las variables, en caso de ser nulo el orden es por nivel de porcentaje}

\item{limite_graf}{Default limite_graf = 1 Recibe valores decimales, los cuales indican el limite del eje X a nivel de procentaje. El 1 equivale al 100 \%.}

\item{width_cats}{Salto de linea que se aplica a las categorias mostradas}

\item{size}{Tamaño de la linea y el punto de la graficamostrada}

\item{size_pct}{Tamaño del porcentaje mostrado}
}
\description{
Graficar lollipop
}
