% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_mapaRegiones}
\alias{graficar_mapaRegiones}
\title{Graficar mapa por categorías}
\usage{
graficar_mapaRegiones(bd, variable, categorica = T)
}
\arguments{
\item{bd}{Base de datos resultado de la función 'calcular_ganadorRegion}

\item{variable}{Variable categórica a graficar}
}
\description{
Graficar mapa por categorías
}
