% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_nube_palabras}
\alias{graficar_nube_palabras}
\title{Graficar nube de palabras}
\usage{
graficar_nube_palabras(bd, max_size)
}
\arguments{
\item{bd}{Base de datos con la estructura generada por calcular_proporciones_nubes}

\item{max_size}{Tamano maximo de las nubes de palabras}
}
\description{
Graficar nube de palabras
}
