% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_sankey}
\alias{graficar_sankey}
\title{Graficar sankey}
\usage{
graficar_sankey(bd, variables, colores, size_text_cat, width_text = 15)
}
\arguments{
\item{bd}{Base de datos procesada con la función \code{\link[encuestar:::analizar_sankey]{encuestar::::analizar_sankey()}}}

\item{variables}{Vector ordenado tipo caracter que contiene los nombres de las llaves de las cuales se va a hacer el cruce}

\item{colores}{Argumento usado por scale_fill_manual y sclae_color_manual}

\item{size_text_cat}{Parametro \link{size} de la funcion \code{\link[treemapify:geom_treemap_subgroup_border]{treemapify::geom_treemap_subgroup_border()}} que modifica el grosor de las lineas que dividen los subgrupos de un bloque}

\item{width_text}{Salto de linea que se aplica a las categorias mostradas}
}
\description{
Graficar sankey
}
\examples{
graficar_sankey(bd = bd_estimacion, size_text_cat = 8)
}
