% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar.R
\name{graficar_sankey}
\alias{graficar_sankey}
\title{Graficar sankey}
\usage{
graficar_sankey(bd, size_text_cat)
}
\arguments{
\item{bd}{Base de datos procesada con la función analizar_sankey}

\item{size_text_cat}{Tamaño del texto mostrado en cada nodo el sankey}
}
\description{
Graficar sankey
}
\examples{
graficar_sankey(bd = bd_estimacion, size_text_cat = 8)
}
