% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Modelo}
\alias{Modelo}
\title{Esta es la clase de Modelo}
\description{
Esta es la clase de Modelo

Esta es la clase de Modelo
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Modelo-new}{\code{Modelo$new()}}
\item \href{#method-Modelo-correspondencia}{\code{Modelo$correspondencia()}}
\item \href{#method-Modelo-componentesPrincipales}{\code{Modelo$componentesPrincipales()}}
\item \href{#method-Modelo-blackBox}{\code{Modelo$blackBox()}}
\item \href{#method-Modelo-clone}{\code{Modelo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-new"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Modelo$new(
  encuesta = NULL,
  diseno = NULL,
  diccionario = NULL,
  tema,
  graficadas = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-correspondencia"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-correspondencia}{}}}
\subsection{Method \code{correspondencia()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Modelo$correspondencia(
  var1,
  var2,
  legenda1 = NULL,
  legenda2 = NULL,
  colores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-componentesPrincipales"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-componentesPrincipales}{}}}
\subsection{Method \code{componentesPrincipales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Modelo$componentesPrincipales(variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-blackBox"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-blackBox}{}}}
\subsection{Method \code{blackBox()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Modelo$blackBox(vars, stimuli)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Modelo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Modelo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Modelo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
