% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Muestra}
\alias{Muestra}
\title{Esta es la clase de muestra}
\description{
Esta es la clase de muestra

Esta es la clase de muestra
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Muestra$new()}}
\item \href{#method-recalcular_fpc}{\code{Muestra$recalcular_fpc()}}
\item \href{#method-extraer_diseño}{\code{Muestra$extraer_diseño()}}
\item \href{#method-clone}{\code{Muestra$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Muestra$new(base)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recalcular_fpc"></a>}}
\if{latex}{\out{\hypertarget{method-recalcular_fpc}{}}}
\subsection{Method \code{recalcular_fpc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Muestra$recalcular_fpc(respuestas)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extraer_diseño"></a>}}
\if{latex}{\out{\hypertarget{method-extraer_diseño}{}}}
\subsection{Method \code{extraer_diseño()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Muestra$extraer_diseño(respuestas, postestratificacion)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Muestra$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
