% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clase_opinometro.R
\name{Opinometro}
\alias{Opinometro}
\title{Clase Opinómetro}
\description{
La clase \code{Opinómetro} gestiona las rutinas necesarias para determinar el diccionario
de procesamiento y la base de datos de respuestas en campo. La clase \code{Opinómetro} es de
importante relevancia pues brinda a la paquetería una fuente de datos e insumos distinta a la que
históricamente se a usado que es \code{SurveyToGo}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id_cuestionarioOpinometro}}{Campo heredado de la clase \code{Encuesta}.}

\item{\code{pool}}{Campo heredado de la clase \code{Encuesta}.}

\item{\code{variables_cuestionario}}{Nombres de las variables que, de acuerdo al diccionario, forman el
conjunto de preguntas de la encuesta}

\item{\code{bd_respuestas_cuestionario}}{\code{\link[=tibble]{tibble()}} que contiene las respuestas serán utilizadas para
construir el diseño muestral de la encuesta.}

\item{\code{diccionario}}{Diccionario (o codebook) del cuestionario de la encuesta. Heredado de la
clase \code{Encuesta} o se provee de forma externa.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Opinometro-new}{\code{Opinometro$new()}}
\item \href{#method-Opinometro-definir_variables}{\code{Opinometro$definir_variables()}}
\item \href{#method-Opinometro-construir_respuestas}{\code{Opinometro$construir_respuestas()}}
\item \href{#method-Opinometro-verificar_variables}{\code{Opinometro$verificar_variables()}}
\item \href{#method-Opinometro-terminar_conexion}{\code{Opinometro$terminar_conexion()}}
\item \href{#method-Opinometro-clone}{\code{Opinometro$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-new"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-new}{}}}
\subsection{Method \code{new()}}{
Establece conexión con la base de datos donde está alojada la
implementación del \code{Opinómetro}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$new(id_cuestionarioOpinometro = NA, pool = NA, diccionario = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_cuestionarioOpinometro}}{Valor entero. Identificador del cuestionario
aplicado en campo generado en la plataforma \code{Opinómetro}.}

\item{\code{pool}}{Objeto tipo \link{pool} generado al conectarse a la base de datos que
almacena las respuestas. Se recomienda utilizar la función \code{\link[=dbPool]{dbPool()}} de la
paquetería \link{pool} para esto.}

\item{\code{diccionario}}{\code{\link[=tibble]{tibble()}} Diccionario (o codebook) del cuestionario de la
encuesta. Aunque la clase \code{Opinometro} reproduce el diccionario a partir del
cuestionario construido en la plataforma \code{Opinometro}, el parametro \code{diccionario}
es aquel que históricamente se ha utilizado para produccion. Es decir, aquel que
está usualmente guardado en un archivo de excel y/o que recibe la clase \code{Encuesta}
en el campo \code{diccionario}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-definir_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-definir_variables}{}}}
\subsection{Method \code{definir_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$definir_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-construir_respuestas"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-construir_respuestas}{}}}
\subsection{Method \code{construir_respuestas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$construir_respuestas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-verificar_variables"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-verificar_variables}{}}}
\subsection{Method \code{verificar_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$verificar_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-terminar_conexion"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-terminar_conexion}{}}}
\subsection{Method \code{terminar_conexion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$terminar_conexion()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Opinometro-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Opinometro-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Opinometro$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
