% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Pregunta}
\alias{Pregunta}
\title{Esta es la clase de pregunta}
\description{
Esta es la clase de pregunta

Esta es la clase de pregunta
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Pregunta-new}{\code{Pregunta$new()}}
\item \href{#method-Pregunta-graficar}{\code{Pregunta$graficar()}}
\item \href{#method-Pregunta-multirespuesta}{\code{Pregunta$multirespuesta()}}
\item \href{#method-Pregunta-regiones_shp}{\code{Pregunta$regiones_shp()}}
\item \href{#method-Pregunta-correspondencia}{\code{Pregunta$correspondencia()}}
\item \href{#method-Pregunta-mapa_ganador}{\code{Pregunta$mapa_ganador()}}
\item \href{#method-Pregunta-mapa_numerico}{\code{Pregunta$mapa_numerico()}}
\item \href{#method-Pregunta-frecuencia_region}{\code{Pregunta$frecuencia_region()}}
\item \href{#method-Pregunta-conocimiento_region}{\code{Pregunta$conocimiento_region()}}
\item \href{#method-Pregunta-saldo_region}{\code{Pregunta$saldo_region()}}
\item \href{#method-Pregunta-resaltar_region}{\code{Pregunta$resaltar_region()}}
\item \href{#method-Pregunta-pclave_region}{\code{Pregunta$pclave_region()}}
\item \href{#method-Pregunta-sankey}{\code{Pregunta$sankey()}}
\item \href{#method-Pregunta-prcomp}{\code{Pregunta$prcomp()}}
\item \href{#method-Pregunta-blackbox_1d}{\code{Pregunta$blackbox_1d()}}
\item \href{#method-Pregunta-morena}{\code{Pregunta$morena()}}
\item \href{#method-Pregunta-cruce_puntos}{\code{Pregunta$cruce_puntos()}}
\item \href{#method-Pregunta-cruce_2vbrechas}{\code{Pregunta$cruce_2vbrechas()}}
\item \href{#method-Pregunta-cruce_multibrechas}{\code{Pregunta$cruce_multibrechas()}}
\item \href{#method-Pregunta-cruce_barras}{\code{Pregunta$cruce_barras()}}
\item \href{#method-Pregunta-cruce_bloque}{\code{Pregunta$cruce_bloque()}}
\item \href{#method-Pregunta-faltantes}{\code{Pregunta$faltantes()}}
\item \href{#method-Pregunta-clone}{\code{Pregunta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$new(encuesta, tema = tema_default)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-graficar"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-graficar}{}}}
\subsection{Method \code{graficar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$graficar(
  llave,
  tipo,
  aspectos = NULL,
  filtro = NULL,
  llave_partido,
  llave_conocimiento,
  llave_opinion,
  llave_xq,
  parametros = list(tit = "", porcentajes_afuera = F, desplazar_porcentajes = 0.01)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-multirespuesta"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-multirespuesta}{}}}
\subsection{Method \code{multirespuesta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$multirespuesta(patron_inicial, tit = "", salto = 100, nota = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-regiones_shp"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-regiones_shp}{}}}
\subsection{Method \code{regiones_shp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$regiones_shp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-correspondencia"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-correspondencia}{}}}
\subsection{Method \code{correspondencia()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$correspondencia(
  var1,
  var2,
  legenda1 = NULL,
  legenda2 = NULL,
  colores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-mapa_ganador"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-mapa_ganador}{}}}
\subsection{Method \code{mapa_ganador()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$mapa_ganador(var, lugar = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-mapa_numerico"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-mapa_numerico}{}}}
\subsection{Method \code{mapa_numerico()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$mapa_numerico(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-frecuencia_region"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-frecuencia_region}{}}}
\subsection{Method \code{frecuencia_region()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$frecuencia_region(variable)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-conocimiento_region"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-conocimiento_region}{}}}
\subsection{Method \code{conocimiento_region()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$conocimiento_region(
  llave_conocimiento,
  candidatos,
  respuesta,
  orden_horizontal
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-saldo_region"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-saldo_region}{}}}
\subsection{Method \code{saldo_region()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$saldo_region(
  llave_opinion = "",
  candidatos,
  ns_nc,
  cat_negativo,
  cat_regular,
  cat_positivo,
  orden_horizontal
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-resaltar_region"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-resaltar_region}{}}}
\subsection{Method \code{resaltar_region()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$resaltar_region(color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-pclave_region"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-pclave_region}{}}}
\subsection{Method \code{pclave_region()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$pclave_region(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-sankey"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-sankey}{}}}
\subsection{Method \code{sankey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$sankey(var1, var2, size_text_cat = 8)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-prcomp"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-prcomp}{}}}
\subsection{Method \code{prcomp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$prcomp(variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-blackbox_1d"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-blackbox_1d}{}}}
\subsection{Method \code{blackbox_1d()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$blackbox_1d(vars, stimuli)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-morena"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-morena}{}}}
\subsection{Method \code{morena()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$morena(personajes, atributos, labels, p)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-cruce_puntos"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-cruce_puntos}{}}}
\subsection{Method \code{cruce_puntos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$cruce_puntos(cruce, variables, vartype = "se", valor_variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-cruce_2vbrechas"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-cruce_2vbrechas}{}}}
\subsection{Method \code{cruce_2vbrechas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$cruce_2vbrechas(
  var1,
  var2_filtro,
  filtro,
  vartype = "cv",
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-cruce_multibrechas"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-cruce_multibrechas}{}}}
\subsection{Method \code{cruce_multibrechas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$cruce_multibrechas(
  por_grupo,
  variables,
  vartype = "cv",
  valor_variables,
  line_rich = FALSE,
  line_linewidth = 2,
  line_hjust = "ymax",
  line_vjust = -0.3
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-cruce_barras"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-cruce_barras}{}}}
\subsection{Method \code{cruce_barras()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$cruce_barras(
  por_grupo,
  variables,
  vartype = "cv",
  valor_variables,
  color,
  filter = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-cruce_bloque"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-cruce_bloque}{}}}
\subsection{Method \code{cruce_bloque()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$cruce_bloque(
  cruce,
  variable,
  vartype = "cv",
  filter = NULL,
  linea_grosor,
  linea_color
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-faltantes"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-faltantes}{}}}
\subsection{Method \code{faltantes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$faltantes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pregunta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Pregunta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pregunta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
