% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opinometro.R
\name{rectificar_respuestasOpinometro}
\alias{rectificar_respuestasOpinometro}
\title{rectificar_respuestasOpinometro}
\usage{
rectificar_respuestasOpinometro(
  bd_respuestas_raw,
  variables_cuestionario,
  var_ubicacion = "gps",
  elim_na_ub = T
)
}
\arguments{
\item{variables_cuestionario}{}
}
\description{
rectificar_respuestasOpinometro
}
