% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_opinometro.R
\name{rectificar_respuestasOpinometro}
\alias{rectificar_respuestasOpinometro}
\title{Title}
\usage{
rectificar_respuestasOpinometro(
  bd_respuestasOpinometro,
  variables_cuestionario
)
}
\arguments{
\item{variables_cuestionario}{}
}
\description{
Title
}
