% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Regiones}
\alias{Regiones}
\title{Esta es la clase de Regiones}
\description{
Esta es la clase de Regiones

Esta es la clase de Regiones
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Regiones-new}{\code{Regiones$new()}}
\item \href{#method-Regiones-mapa_ganador}{\code{Regiones$mapa_ganador()}}
\item \href{#method-Regiones-mapa_degradadoNumerico}{\code{Regiones$mapa_degradadoNumerico()}}
\item \href{#method-Regiones-heatmap_conocimiento}{\code{Regiones$heatmap_conocimiento()}}
\item \href{#method-Regiones-heatmap_saldoOpinion}{\code{Regiones$heatmap_saldoOpinion()}}
\item \href{#method-Regiones-mapa_resaltarRegion}{\code{Regiones$mapa_resaltarRegion()}}
\item \href{#method-Regiones-clone}{\code{Regiones$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-new"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$new(diseno, diccionario = NULL, shp = NULL, tema)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-mapa_ganador"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-mapa_ganador}{}}}
\subsection{Method \code{mapa_ganador()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$mapa_ganador(variable, lugar = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-mapa_degradadoNumerico"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-mapa_degradadoNumerico}{}}}
\subsection{Method \code{mapa_degradadoNumerico()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$mapa_degradadoNumerico(variable)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-heatmap_conocimiento"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-heatmap_conocimiento}{}}}
\subsection{Method \code{heatmap_conocimiento()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$heatmap_conocimiento(
  llave_conocimiento,
  candidatos,
  respuesta,
  orden_horizontal
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-heatmap_saldoOpinion"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-heatmap_saldoOpinion}{}}}
\subsection{Method \code{heatmap_saldoOpinion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$heatmap_saldoOpinion(
  llave_opinion = "",
  candidatos,
  ns_nc,
  cat_negativo,
  cat_regular,
  cat_positivo,
  orden_horizontal
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-mapa_resaltarRegion"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-mapa_resaltarRegion}{}}}
\subsection{Method \code{mapa_resaltarRegion()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$mapa_resaltarRegion(region, color)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Regiones-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Regiones-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Regiones$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
