% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Respuestas}
\alias{Respuestas}
\title{Esta es la clase Respuestas}
\description{
Esta es la clase Respuestas

Esta es la clase Respuestas
}
\section{Super class}{
\code{\link[encuestar:Encuesta]{encuestar::Encuesta}} -> \code{Respuestas}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Respuestas$new()}}
\item \href{#method-eliminar_auditoria_telefonica}{\code{Respuestas$eliminar_auditoria_telefonica()}}
\item \href{#method-eliminar_faltantes_diseño}{\code{Respuestas$eliminar_faltantes_diseño()}}
\item \href{#method-clone}{\code{Respuestas$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crear respuesta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$new(base = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Base de datos de respuestas.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_auditoria_telefonica"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_auditoria_telefonica}{}}}
\subsection{Method \code{eliminar_auditoria_telefonica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_auditoria_telefonica(auditoria_telefonica)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_faltantes_diseño"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_faltantes_diseño}{}}}
\subsection{Method \code{eliminar_faltantes_diseño()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_faltantes_diseño()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
