% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_surveytogo.R
\name{Respuestas}
\alias{Respuestas}
\title{Esta es la clase Respuestas}
\description{
Esta es la clase Respuestas

Esta es la clase Respuestas
}
\section{Super class}{
\code{\link[encuestar:Encuesta]{encuestar::Encuesta}} -> \code{Respuestas}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Respuestas$new()}}
\item \href{#method-eliminar_auditoria_telefonica}{\code{Respuestas$eliminar_auditoria_telefonica()}}
\item \href{#method-eliminar_falta_coordenadas}{\code{Respuestas$eliminar_falta_coordenadas()}}
\item \href{#method-correccion_cluster}{\code{Respuestas$correccion_cluster()}}
\item \href{#method-eliminar_fuera_muestra}{\code{Respuestas$eliminar_fuera_muestra()}}
\item \href{#method-eliminar_faltantes_diseno}{\code{Respuestas$eliminar_faltantes_diseno()}}
\item \href{#method-vars_diseno}{\code{Respuestas$vars_diseno()}}
\item \href{#method-clone}{\code{Respuestas$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="encuestar" data-topic="Encuesta" data-id="exportar_entregable">}\href{../../encuestar/html/Encuesta.html#method-exportar_entregable}{\code{encuestar::Encuesta$exportar_entregable()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Crear respuesta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$new(
  base,
  auditoria_telefonica,
  muestra,
  shp,
  mantener,
  diccionario,
  nivel,
  var_n
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Base de datos de respuestas.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_auditoria_telefonica"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_auditoria_telefonica}{}}}
\subsection{Method \code{eliminar_auditoria_telefonica()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_auditoria_telefonica(auditoria_telefonica)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_falta_coordenadas"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_falta_coordenadas}{}}}
\subsection{Method \code{eliminar_falta_coordenadas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_falta_coordenadas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correccion_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-correccion_cluster}{}}}
\subsection{Method \code{correccion_cluster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$correccion_cluster(base, shp, mantener, nivel, var_n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_fuera_muestra"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_fuera_muestra}{}}}
\subsection{Method \code{eliminar_fuera_muestra()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_fuera_muestra(respuestas, muestra, nivel, var_n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eliminar_faltantes_diseno"></a>}}
\if{latex}{\out{\hypertarget{method-eliminar_faltantes_diseno}{}}}
\subsection{Method \code{eliminar_faltantes_diseno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$eliminar_faltantes_diseno()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vars_diseno"></a>}}
\if{latex}{\out{\hypertarget{method-vars_diseno}{}}}
\subsection{Method \code{vars_diseno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$vars_diseno(muestra, var_n, tipo_encuesta)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Respuestas$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
