% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clase_resultados.R
\name{Resultados}
\alias{Resultados}
\title{Esta es la clase de Resultados}
\description{
Esta es la clase de Resultados

Esta es la clase de Resultados
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Resultados-new}{\code{Resultados$new()}}
\item \href{#method-Resultados-faltantes}{\code{Resultados$faltantes()}}
\item \href{#method-Resultados-clone}{\code{Resultados$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resultados-new"></a>}}
\if{latex}{\out{\hypertarget{method-Resultados-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resultados$new(encuesta, diseno, diccionario, tema = tema_morant())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resultados-faltantes"></a>}}
\if{latex}{\out{\hypertarget{method-Resultados-faltantes}{}}}
\subsection{Method \code{faltantes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resultados$faltantes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Resultados-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Resultados-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Resultados$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
